🌱 Split agents.py router into modules under 1000 lines
Splits the 5091-line app/routers/agents.py by functional area. agents.py
keeps the router, the read/delete endpoints and re-exports every moved
name, so existing importers and tests are unchanged.

Removes the dead GET /agents/migration/migratable route: it was
registered after GET /{namespace}/{name}, which matches the same two
path segments, so requests already resolved to get_agent. Only the
decorator is gone -- migrate_all_agents still calls the function.

All 83 moved definitions are textually identical to the original.
