+++ rossoctl/backend/app/routers/agents.py
@@ -1,674 +1,177 @@
-# pylint: disable=too-many-lines
 # Copyright 2025 IBM Corp.
 # Licensed under the Apache License, Version 2.0
 
 """
 Agent API endpoints.
+
+This module owns the agent read/delete endpoints and composes the rest of the
+Agent API from focused submodules:
+
+* ``agents_models``     -- request/response models
+* ``agents_status``     -- workload readiness/status derivation
+* ``agents_manifests``  -- Deployment/StatefulSet/Job/Sandbox/Service builders
+* ``agents_skills``     -- linked-skill wiring and Shipwright build manifests
+* ``agents_authbridge`` -- AuthBridge ConfigMaps/RBAC and identity endpoints
+* ``agents_migration``  -- legacy Agent CRD -> workload migration
+* ``agents_shipwright`` -- Shipwright build/buildrun endpoints
+* ``agents_create``     -- agent creation
+* ``agents_finalize``   -- finalization of source-built agents
+* ``agents_env``        -- .env parsing/fetching helpers
+
+Names those submodules define are re-exported here (see ``__all__``) because
+other modules and the test suite import them from ``app.routers.agents``.
 """
 
-import json
 import logging
-import re
-import socket
-import ipaddress
-from datetime import datetime, timezone
-from typing import Any, Dict, List, Literal, Optional, Tuple
-from urllib.parse import urlparse
+from typing import Any
 
-import httpx
-import yaml
-from fastapi import APIRouter, Body, Depends, HTTPException, Query
-import kubernetes.client
+from fastapi import APIRouter, Depends, HTTPException, Query
 from kubernetes.client import ApiException
-from pydantic import BaseModel, Field, field_validator, model_validator
 
 from app.core.auth import ROLE_OPERATOR, ROLE_VIEWER, require_roles
-from app.utils.routes import get_agent_url
+from app.core.config import settings
 from app.core.constants import (
+    AGENTRUNTIMES_PLURAL,
+    AGENTS_PLURAL,
     CRD_GROUP,
     CRD_VERSION,
-    AGENTS_PLURAL,
-    AGENTRUNTIMES_PLURAL,
-    ROSSOCTL_TYPE_LABEL,
-    PROTOCOL_LABEL_PREFIX,
-    ROSSOCTL_FRAMEWORK_LABEL,
-    ROSSOCTL_INJECT_LABEL,
-    ROSSOCTL_WORKLOAD_TYPE_LABEL,
-    ROSSOCTL_DESCRIPTION_ANNOTATION,
-    APP_KUBERNETES_IO_CREATED_BY,
-    APP_KUBERNETES_IO_NAME,
-    APP_KUBERNETES_IO_MANAGED_BY,
-    APP_KUBERNETES_IO_COMPONENT,
-    ROSSOCTL_UI_CREATOR_LABEL,
-    ROSSOCTL_OPERATOR_LABEL_NAME,
     RESOURCE_TYPE_AGENT,
-    DEFAULT_IN_CLUSTER_PORT,
-    DEFAULT_OFF_CLUSTER_PORT,
-    DEFAULT_IMAGE_POLICY,
-    DEFAULT_RESOURCE_LIMITS,
-    DEFAULT_RESOURCE_REQUESTS,
-    DEFAULT_ENV_VARS,
-    AGENT_ENDPOINT,
-    AGENT_SKILLS_ANNOTATION,
-    AGENT_SKILLS_MOUNT_ROOT,
-    SKILL_TYPE_LABEL,
-    SKILL_TYPE_VALUE,
-    SKILL_DISPLAY_NAME_ANNOTATION,
-    SKILL_DESCRIPTION_ANNOTATION,
-    # Shipwright constants
+    ROSSOCTL_DESCRIPTION_ANNOTATION,
+    ROSSOCTL_TYPE_LABEL,
+    ROSSOCTL_WORKLOAD_TYPE_LABEL,
+    SHIPWRIGHT_BUILDRUNS_PLURAL,
+    SHIPWRIGHT_BUILDS_PLURAL,
     SHIPWRIGHT_CRD_GROUP,
     SHIPWRIGHT_CRD_VERSION,
-    SHIPWRIGHT_BUILDS_PLURAL,
-    SHIPWRIGHT_BUILDRUNS_PLURAL,
-    SHIPWRIGHT_CLUSTER_BUILD_STRATEGIES_PLURAL,
-    DEFAULT_INTERNAL_REGISTRY,
-    # Workload type constants
     WORKLOAD_TYPE_DEPLOYMENT,
-    WORKLOAD_TYPE_STATEFULSET,
     WORKLOAD_TYPE_JOB,
     WORKLOAD_TYPE_SANDBOX,
-    AGENT_SANDBOX_CRD_GROUP,
-    AGENT_SANDBOX_CRD_VERSION,
-    SUPPORTED_WORKLOAD_TYPES,
-    # Migration constants (Phase 4)
-    MIGRATION_SOURCE_ANNOTATION,
-    MIGRATION_TIMESTAMP_ANNOTATION,
-    # SPIRE identity constants
-    ROSSOCTL_SPIRE_LABEL,
-    ROSSOCTL_SPIRE_ENABLED_VALUE,
-    # Port exclusion annotations
-    ROSSOCTL_OUTBOUND_PORTS_EXCLUDE,
-    ROSSOCTL_INBOUND_PORTS_EXCLUDE,
-    # AuthBridge ConfigMap defaults
-    DEFAULT_KEYCLOAK_INTERNAL_URL,
-    DEFAULT_KEYCLOAK_REALM,
-    DEFAULT_SPIFFE_HELPER_CONF,
-    DEFAULT_ENVOY_YAML,
-    # External skill registry constants
-    SKILL_SOURCE_LABEL,
-    SKILL_SOURCE_EXTERNAL,
-    SKILL_REGISTRY_TYPE_LABEL,
-    SKILL_REGISTRY_URL_ANNOTATION,
-    SKILL_REGISTRY_SKILL_NAME_ANNOTATION,
-    SKILL_REGISTRY_SKILL_VERSION_ANNOTATION,
-    SKILL_FETCHER_SCRIPTS_CM,
-    SKILL_FETCHER_IMAGE,
+    WORKLOAD_TYPE_STATEFULSET,
 )
-from app.core.config import settings
 from app.models.responses import (
-    AgentSummary,
     AgentListResponse,
-    ResourceLabels,
+    AgentSummary,
     DeleteResponse,
+    ResourceLabels,
 )
-from app.services.agent_env_defaults import apply_agent_import_defaults
+from app.models.shipwright import ShipwrightBuildConfig
 from app.services.kubernetes import KubernetesService, get_kubernetes_service
-from app.routers.skills import _sanitize_k8s_name
-from app.utils.routes import (
-    create_route_for_agent_or_tool,
-    detect_platform,
-    rollback_workload_resources,
-    route_exists,
-    sanitize_log,
-    select_route_port,
+from app.services.shipwright import extract_buildrun_info, get_latest_buildrun
+from app.services.shipwright_builds import collect_rossoctl_shipwright_builds
+from app.utils.routes import route_exists, sanitize_log
+
+# --- Re-exports -------------------------------------------------------------
+# Imported for backwards compatibility: other modules and tests import these
+# from `app.routers.agents`. See __all__ at the bottom of this module.
+from app.routers.agents_authbridge import (
+    authbridge_router,
+    _build_authbridge_runtime_yaml,
+    _build_authbridge_runtime_yaml_fallback,
+    _ensure_authbridge_configmaps,
+    _ensure_authbridge_scc_rolebinding,
+    _ensure_authproxy_routes,
+    _fetch_authbridge_json,
+    _get_authbridge_runtime_yaml,
+    _get_service_endpoints,
 )
-from app.models.shipwright import (
-    ResourceType,
-    ShipwrightBuildConfig,
-    BuildSourceConfig,
-    BuildOutputConfig,
-    BuildStatusCondition,
-    ClusterBuildStrategyInfo,
-    ClusterBuildStrategiesResponse,
-    ShipwrightBuildListResponse,
-    ShipwrightBuildStatusResponse,
-    ShipwrightBuildRunStatusResponse,
-    ResourceConfigFromBuild,
-)
-from app.services.shipwright import (
-    build_shipwright_build_manifest,
-    build_shipwright_buildrun_manifest,
-    extract_resource_config_from_build,
-    get_latest_buildrun,
-    extract_buildrun_info,
-    resolve_clone_secret,
-)
-from app.services.shipwright_builds import (
-    cleanup_existing_build,
-    collect_rossoctl_shipwright_builds,
+from app.routers.agents_create import create_agent
+from app.routers.agents_env import (
+    env_router,
+    BLOCKED_IP_RANGES,
+    fetch_env_from_url,
+    is_ip_blocked,
+    parse_env_file,
 )
 
-
-class OutboundRoute(BaseModel):
-    """A single outbound token exchange route for authproxy-routes ConfigMap."""
-
-    host: str = Field(..., min_length=1)
-    target_audience: str = Field(..., min_length=1)
-    token_scopes: str = "openid"
-
-
-class SecretKeyRef(BaseModel):
-    """Reference to a key in a Secret."""
-
-    name: str
-    key: str
-
-
-class ConfigMapKeyRef(BaseModel):
-    """Reference to a key in a ConfigMap."""
-
-    name: str
-    key: str
-
-
-class EnvVarSource(BaseModel):
-    """Source for environment variable value."""
-
-    secretKeyRef: Optional[SecretKeyRef] = None
-    configMapKeyRef: Optional[ConfigMapKeyRef] = None
-
-
-class EnvVar(BaseModel):
-    """Environment variable with support for direct values and references."""
-
-    name: str
-    value: Optional[str] = None
-    valueFrom: Optional[EnvVarSource] = None
-
-    @field_validator("name")
-    @classmethod
-    def validate_env_var_name(cls, v: str) -> str:
-        """Validate environment variable name according to Kubernetes rules.
-
-        Valid env var names must:
-        - Contain only letters (A-Z, a-z), digits (0-9), and underscores (_)
-        - Not start with a digit
-        """
-        if not v:
-            raise ValueError("Environment variable name cannot be empty")
-
-        # Kubernetes env var name pattern: must start with letter or underscore,
-        # followed by any combination of letters, digits, or underscores
-        pattern = r"^[A-Za-z_][A-Za-z0-9_]*$"
-
-        if not re.match(pattern, v):
-            raise ValueError(
-                f"Invalid environment variable name '{v}'. "
-                "Name must start with a letter or underscore and contain only "
-                "letters, digits, and underscores (e.g., MY_VAR, API_KEY, var123)."
-            )
-
-        return v
-
-    @field_validator("valueFrom")
-    @classmethod
-    def check_value_or_value_from(cls, v, info):
-        """Ensure either value or valueFrom is provided, but not both."""
-        values = info.data
-        has_value = values.get("value") is not None
-        has_value_from = v is not None
-
-        if not has_value and not has_value_from:
-            raise ValueError("Either value or valueFrom must be provided")
-        if has_value and has_value_from:
-            raise ValueError("Cannot specify both value and valueFrom")
-
-        return v
-
-
-class ServicePort(BaseModel):
-    """Service port configuration."""
-
-    name: str = "http"
-    port: int = 8080
-    targetPort: int = 8000
-    protocol: str = "TCP"
-
-
-class PersistentStorageConfig(BaseModel):
-    """Persistent storage configuration for Sandbox and StatefulSet agents."""
-
-    enabled: bool = False
-    size: str = "1Gi"
-
-
-class ContextAttachment(BaseModel):
-    """A named Context Service resource mounted into the agent."""
-
-    name: str
-    mountPath: str = "/workspace"
-    readOnly: bool = False
-
-
-class CreateAgentRequest(BaseModel):
-    """Request to create a new agent."""
-
-    name: str
-    namespace: str
-    protocol: str = "a2a"
-    framework: str = "LangGraph"
-    envVars: Optional[List[EnvVar]] = None
-    skills: Optional[List[str]] = None
-    # Optional MCP tool link — when agent import defaults are enabled, MCP_URL is injected
-    mcpToolName: Optional[str] = None
-    # LLM preset: openai, ollama, openrouter (used with agent import defaults)
-    llmPreset: Optional[str] = None
-    llmModel: Optional[str] = None
-
-    # Workload type: 'deployment', 'statefulset', or 'job'
-    workloadType: str = WORKLOAD_TYPE_DEPLOYMENT
-
-    # Deployment method: 'source' (build from git) or 'image' (use existing image)
-    deploymentMethod: str = "source"
-
-    # Build from source fields
-    gitUrl: str = ""
-    gitPath: str = ""
-    gitBranch: str = "main"
-    imageTag: str = "v0.0.1"
-    registryUrl: Optional[str] = None
-    registrySecret: Optional[str] = None
-    startCommand: Optional[str] = None
-
-    # Deploy from existing image fields
-    containerImage: Optional[str] = None
-    imagePullSecret: Optional[str] = None
-
-    # Pod configuration
-    servicePorts: Optional[List[ServicePort]] = None
-
-    # HTTPRoute/Route creation
-    createHttpRoute: bool = False
-
-    # AuthBridge sidecar injection (default enabled for agents)
-    authBridgeEnabled: bool = True
-    # SPIRE identity (gates spiffe-helper inside the combined authbridge container)
-    spireEnabled: bool = False
-
-    # Per-workload AuthBridge mode override. Maps to
-    # AgentRuntime.Spec.AuthBridgeMode; when None the operator falls
-    # back through namespace ConfigMap → cluster default
-    # (proxy-sidecar). The lite/waypoint shapes are accepted by the
-    # operator but not surfaced through the UI today.
-    authBridgeMode: Optional[Literal["proxy-sidecar", "envoy-sidecar", "lite", "waypoint"]] = None
-
-    # Per-workload mTLS posture between AuthBridge sidecars. Maps to
-    # AgentRuntime.Spec.MTLSMode. The rossoctl UI sends an explicit
-    # value (default "disabled") so users always see what they get;
-    # this means UI-created agents opt out of any namespace-level
-    # mtls.mode setting (CR-pin semantic in the operator). The
-    # operator's validating webhook rejects the envoy-sidecar +
-    # non-disabled combo because Envoy SDS isn't currently configured
-    # by the rossoctl envoy-config — we mirror that check below as a
-    # model_validator so the form gets a fast 422 instead of a
-    # webhook denial after the manifest is built.
-    mtlsMode: Optional[Literal["disabled", "permissive", "strict"]] = None
-
-    # Per-workload TLS bridge: decrypt the agent's outbound HTTPS so AuthBridge's
-    # pipeline can inspect it. Maps to AgentRuntime.Spec.TLSBridgeMode (enabled
-    # when True; left unset → operator default "disabled"). Like mtlsMode, it's a
-    # plain per-agent field — no operator feature gate and no UI feature flag; the
-    # import-form checkbox shows whenever AuthBridge is enabled. The bridge lives
-    # in the Go forward proxy, so it requires proxy-sidecar/lite mode — the
-    # validator below mirrors the operator webhook's reject of envoy-sidecar. It
-    # also needs cert-manager and an operator build that supports the bridge.
-    tlsBridgeEnabled: bool = False
-
-    # Port exclusion annotations
-    outboundPortsExclude: Optional[str] = None
-    inboundPortsExclude: Optional[str] = None
-
-    # AuthBridge config overrides (authbridge-config ConfigMap)
-    defaultOutboundPolicy: Optional[Literal["passthrough", "exchange"]] = None
-
-    # Outbound routing rules (authproxy-routes ConfigMap)
-    outboundRoutes: Optional[List["OutboundRoute"]] = None
-
-    # Persistent storage (for Sandbox and StatefulSet workloads)
-    persistentStorage: Optional[PersistentStorageConfig] = None
-
-    # Named Context Service resources. Requires CONTEXT_SERVICE_URL.
-    contexts: Optional[List[ContextAttachment]] = None
-
-    # Shipwright build configuration
-    shipwrightConfig: Optional[ShipwrightBuildConfig] = None
-
-    # Optional per-agent overrides for container resource limits/requests
-    # (falls back to DEFAULT_RESOURCE_LIMITS / DEFAULT_RESOURCE_REQUESTS).
-    #
-    # Note: The keys and quantity strings are not validated. This is
-    # deliberate: validating them here would mean testing Kubernetes
-    # shapes (including extended resources such as nvidia.com/gpu),
-    # a complex validation path that would still have to be kept in sync
-    # with the API server.
-    k8sResourceLimits: Optional[Dict[str, str]] = None
-    k8sResourceRequests: Optional[Dict[str, str]] = None
-
-    @field_validator("workloadType")
-    @classmethod
-    def validate_workload_type(cls, v: str) -> str:
-        """Validate that workload type is supported."""
-        if v not in SUPPORTED_WORKLOAD_TYPES:
-            raise ValueError(
-                f"Unsupported workload type: {v}. "
-                f"Supported types: {', '.join(SUPPORTED_WORKLOAD_TYPES)}"
-            )
-        return v
-
-    @model_validator(mode="after")
-    def _check_mtls_compatible_with_mode(self) -> "CreateAgentRequest":
-        """Hook for cross-field rejections of authBridgeMode +
-        mtlsMode combinations the operator's AgentRuntime validating
-        webhook would also reject.
-
-        envoy-sidecar + non-disabled mtlsMode used to be rejected here
-        as defense-in-depth in front of the operator's webhook gate.
-        Both have been lifted now that the operator + extensions
-        support the full matrix (rossoctl-operator#381,
-        cortex#441), so today there are no rejected
-        combinations.
-
-        TODO(future-incompatibility): re-enable cross-field rejections
-        here when a new authBridgeMode (e.g. waypoint, sidecarless)
-        lands that needs different mTLS semantics. The function
-        intentionally stays as a single grep-target so the rejection
-        can land here instead of getting scattered across the request
-        model. Mirrors the operator's checkMTLSCompatibleWithMode
-        pattern in agentruntime_webhook.go.
-
-        SPIRE-vs-mTLS coupling is intentionally NOT enforced here:
-        rossoctl-operator's pod_mutator auto-enables SPIRE when
-        mtlsMode != disabled (pod_mutator.go:288-302), so a request
-        with mtlsMode=strict + spireEnabled=false is handled at the
-        operator data-plane layer rather than rejected at the API
-        boundary. The UI form still locks the mTLS dropdown when
-        SPIRE is off as a UX hint, but a non-UI client (CLI, direct
-        API call) submitting that combination is valid and the
-        operator turns SPIRE on for them.
-        """
-        return self
-
-    @model_validator(mode="after")
-    def _check_tlsbridge_compatible_with_mode(self) -> "CreateAgentRequest":
-        """Reject tlsBridgeEnabled with an authBridgeMode that can't host the
-        bridge, mirroring the operator's checkTLSBridgeCompatibleWithMode
-        (agentruntime_webhook.go). The TLS bridge lives in the Go forward proxy,
-        which only exists in proxy-sidecar / lite. Uses the same ALLOWLIST as the
-        operator (empty → defaults to proxy-sidecar) rather than a denylist, so a
-        future authBridgeMode can't slip past this fast-422 and only get rejected
-        at the webhook.
-        """
-        allowed = (None, "", "proxy-sidecar", "lite")
-        if self.tlsBridgeEnabled and self.authBridgeMode not in allowed:
-            raise ValueError(
-                "tlsBridgeEnabled requires authBridgeMode proxy-sidecar or lite "
-                f"(the TLS bridge lives in the Go forward proxy); got {self.authBridgeMode!r}"
-            )
-        return self
-
-
-class CreateAgentResponse(BaseModel):
-    """Response after creating an agent."""
-
-    success: bool
-    name: str
-    namespace: str
-    message: str
-
-
-class AgentShipwrightBuildInfoResponse(BaseModel):  # pylint: disable=too-many-instance-attributes
-    """Full Shipwright Build information for agents.
-
-    This is an agent-specific wrapper that includes agentConfig for backwards compatibility.
-    """
-
-    # Build info
-    name: str
-    namespace: str
-    buildRegistered: bool
-    buildReason: Optional[str] = None
-    buildMessage: Optional[str] = None
-    outputImage: str
-    strategy: str
-    gitUrl: str
-    gitRevision: str
-    contextDir: str
-
-    # Latest BuildRun info (if any)
-    hasBuildRun: bool = False
-    buildRunName: Optional[str] = None
-    buildRunPhase: Optional[str] = None  # Pending, Running, Succeeded, Failed
-    buildRunStartTime: Optional[str] = None
-    buildRunCompletionTime: Optional[str] = None
-    buildRunOutputImage: Optional[str] = None
-    buildRunOutputDigest: Optional[str] = None
-    buildRunFailureMessage: Optional[str] = None
-
-    # Agent configuration from annotations (agent-specific)
-    agentConfig: Optional[ResourceConfigFromBuild] = None
-
-
-# Migration Models (Phase 4: Agent CRD to Deployment migration)
-
-
-class MigrateAgentRequest(BaseModel):
-    """Request to migrate an Agent CRD to a Deployment."""
-
-    delete_old: bool = False  # Whether to delete the Agent CRD after successful migration
-
-
-class MigrateAgentResponse(BaseModel):
-    """Response after migrating an agent."""
-
-    success: bool
-    migrated: bool
-    name: str
-    namespace: str
-    message: str
-    deployment_created: bool = False
-    service_created: bool = False
-    agent_crd_deleted: bool = False
-
-
-class MigratableAgentInfo(BaseModel):
-    """Information about an agent that can be migrated."""
-
-    name: str
-    namespace: str
-    status: str
-    has_deployment: bool  # True if a Deployment already exists with same name
-    labels: Dict[str, str]
-    description: Optional[str] = None
-
-
-class ListMigratableAgentsResponse(BaseModel):
-    """Response containing list of agents that can be migrated."""
-
-    agents: List[MigratableAgentInfo]
-    total: int
-    already_migrated: int  # Count of agents that already have Deployments
+# `finalize_shipwright_build` is re-exported here because
+# app/services/reconciliation.py imports it from `app.routers.agents` inside a
+# function body, and tests patch the string target
+# "app.routers.agents.finalize_shipwright_build". That late import is what makes
+# the patch effective -- do not promote it to a module-level import there.
+from app.routers.agents_finalize import finalize_router, finalize_shipwright_build
+from app.routers.agents_manifests import (
+    build_container_resources,
+    _agentruntime_supported_workload,
+    _build_agentruntime_manifest,
+    _build_common_annotations,
+    _build_common_labels,
+    _build_deployment_manifest,
+    _build_env_vars,
+    _build_job_manifest,
+    _build_sandbox_manifest,
+    _build_selector_labels,
+    _build_service_manifest,
+    _build_statefulset_manifest,
+    _create_or_replace_service,
+    _ensure_agentruntime,
+    _resolve_context_mounts,
+)
+from app.routers.agents_migration import (
+    migration_router,
+    _build_deployment_from_agent_crd,
+    _build_service_from_agent_crd,
+    list_migratable_agents,
+    migrate_agent,
+    migrate_all_agents,
+)
+from app.routers.agents_models import (
+    AgentShipwrightBuildInfoResponse,
+    ConfigMapKeyRef,
+    ContextAttachment,
+    CreateAgentRequest,
+    CreateAgentResponse,
+    EnvVar,
+    EnvVarSource,
+    FetchEnvUrlRequest,
+    FetchEnvUrlResponse,
+    FinalizeShipwrightBuildRequest,
+    ListMigratableAgentsResponse,
+    MigratableAgentInfo,
+    MigrateAgentRequest,
+    MigrateAgentResponse,
+    OutboundRoute,
+    ParseEnvRequest,
+    ParseEnvResponse,
+    PersistentStorageConfig,
+    SecretKeyRef,
+    ServicePort,
+)
+from app.routers.agents_shipwright import (
+    shipwright_router,
+    get_shipwright_build_info,
+    get_shipwright_build_status,
+    get_shipwright_buildrun_status,
+    list_agent_shipwright_builds,
+    list_build_strategies,
+    trigger_shipwright_buildrun,
+)
+from app.routers.agents_skills import (
+    _build_agent_shipwright_build_manifest,
+    _build_agent_shipwright_buildrun_manifest,
+    _build_fetcher_scripts_data,
+    _ensure_card_unsigned_configmap,
+    _ensure_fetcher_scripts_cm,
+    _get_external_skill_data,
+    _get_linked_skill_mounts,
+    _is_skill_external,
+    _load_agent_skill_summaries,
+)
+from app.routers.agents_status import (
+    _extract_labels,
+    _format_timestamp,
+    _get_deployment_description,
+    _get_job_description,
+    _get_job_status,
+    _get_sandbox_description,
+    _get_statefulset_description,
+    _is_deployment_ready,
+    _is_sandbox_ready,
+    _is_statefulset_ready,
+)
 
 
 logger = logging.getLogger(__name__)
 router = APIRouter(prefix="/agents", tags=["agents"])
-
-
-def _is_deployment_ready(resource_data: dict) -> str:
-    """Check if a Kubernetes Deployment is ready based on status.
-
-    For Deployments, checks:
-    1. conditions array for type="Available" with status="True"
-    2. replicas vs readyReplicas count
-
-    Also maintains backward compatibility with Agent CRD status format.
-    """
-    status = resource_data.get("status", {})
-    conditions = status.get("conditions") or []
-
-    # Check for Kubernetes Deployment conditions (type=Available)
-    for condition in conditions:
-        cond_type = condition.get("type")
-        cond_status = condition.get("status")
-
-        # Kubernetes Deployment uses "Available" condition
-        if cond_type == "Available" and cond_status == "True":
-            return "Ready"
-
-        # Agent CRD uses "Ready" condition (backward compatibility)
-        if cond_type == "Ready" and cond_status == "True":
-            return "Ready"
-
-    # Check replica counts for Deployments
-    replicas = status.get("replicas") or 0
-    ready_replicas = status.get("ready_replicas") or status.get("readyReplicas", 0)
-    if 0 < replicas <= ready_replicas:
-        return "Ready"
-
-    # Fallback: check deploymentStatus.phase for older Agent CRD versions
-    deployment_status = status.get("deploymentStatus", {})
-    phase = deployment_status.get("phase", "")
-    if phase in ("Ready", "Running"):
-        return "Ready"
-
-    return "Not Ready"
-
-
-def _get_deployment_description(deployment: dict) -> str:
-    """Extract description from Deployment annotations."""
-    annotations = deployment.get("metadata", {}).get("annotations", {})
-    return annotations.get(
-        ROSSOCTL_DESCRIPTION_ANNOTATION,
-        annotations.get("description", "No description"),
-    )
-
-
-def _is_statefulset_ready(resource_data: dict) -> str:
-    """Check if a Kubernetes StatefulSet is ready based on status."""
-    status = resource_data.get("status", {})
-
-    # Check replica counts for StatefulSets
-    replicas = status.get("replicas") or 0
-    ready_replicas = status.get("ready_replicas") or status.get("readyReplicas", 0)
-
-    if replicas == 0:
-        return "Not Ready"
-    if ready_replicas >= replicas:
-        return "Ready"
-    if ready_replicas > 0:
-        return "Progressing"
-    return "Not Ready"
-
-
-def _get_statefulset_description(statefulset: dict) -> str:
-    """Extract description from StatefulSet annotations."""
-    annotations = statefulset.get("metadata", {}).get("annotations", {})
-    return annotations.get(
-        ROSSOCTL_DESCRIPTION_ANNOTATION,
-        annotations.get("description", "No description"),
-    )
-
-
-def _get_job_status(job: dict) -> str:
-    """Get the status of a Kubernetes Job.
-
-    Returns status values consistent with Deployments and StatefulSets:
-    - "Ready": Job completed successfully (equivalent to Job condition "Complete")
-    - "Failed": Job failed (equivalent to Job condition "Failed")
-    - "Progressing": Job is actively running (has active pods)
-    - "Not Ready": Job is pending/not yet started
-
-    This mapping ensures UI consistency across all workload types.
-    """
-    status = job.get("status", {})
-    conditions = status.get("conditions") or []
-
-    # Check conditions for completed or failed
-    for condition in conditions:
-        cond_type = condition.get("type")
-        cond_status = condition.get("status")
-
-        if cond_type == "Complete" and cond_status == "True":
-            return "Ready"  # Job completed successfully
-        if cond_type == "Failed" and cond_status == "True":
-            return "Failed"
-
-    # Check active/succeeded/failed counts
-    active = status.get("active") or 0
-    succeeded = status.get("succeeded") or 0
-    failed = status.get("failed") or 0
-
-    if succeeded > 0:
-        return "Ready"  # Job completed successfully
-    if failed > 0:
-        return "Failed"
-    if active > 0:
-        return "Progressing"  # Job is actively running
-    return "Not Ready"  # Job pending/not started
-
-
-def _get_job_description(job: dict) -> str:
-    """Extract description from Job annotations."""
-    annotations = job.get("metadata", {}).get("annotations", {})
-    return annotations.get(
-        ROSSOCTL_DESCRIPTION_ANNOTATION,
-        annotations.get("description", "No description"),
-    )
-
-
-def _is_sandbox_ready(sandbox: dict) -> str:
-    """Check if a Sandbox is ready by examining its status conditions."""
-    status = sandbox.get("status", {})
-    conditions = status.get("conditions", [])
-    for cond in conditions:
-        if cond.get("type") == "Ready":
-            if cond.get("status") == "True":
-                return "Ready"
-            return "Not Ready"
-    return "Pending"
-
-
-def _get_sandbox_description(sandbox: dict) -> str:
-    """Extract description from a Sandbox resource."""
-    metadata = sandbox.get("metadata", {})
-    annotations = metadata.get("annotations", {})
-    return annotations.get(ROSSOCTL_DESCRIPTION_ANNOTATION, "No description")
-
-
-def _format_timestamp(timestamp) -> Optional[str]:
-    """Convert a timestamp to ISO format string.
-
-    The Kubernetes Python client returns datetime objects for timestamp fields,
-    but our Pydantic models expect strings.
-    """
-    if timestamp is None:
-        return None
-    if isinstance(timestamp, str):
-        return timestamp
-    # Handle datetime objects from K8s Python client
-    if hasattr(timestamp, "isoformat"):
-        return timestamp.isoformat()
-    return str(timestamp)
-
-
-def _extract_labels(labels: dict) -> ResourceLabels:
-    """Extract rossoctl labels from Kubernetes labels."""
-    # Extract protocols from protocol.rossoctl.io/<name> prefix labels.
-    protocols = [
-        k[len(PROTOCOL_LABEL_PREFIX) :]
-        for k in labels
-        if k.startswith(PROTOCOL_LABEL_PREFIX) and len(k) > len(PROTOCOL_LABEL_PREFIX)
-    ]
-    # Fall back to deprecated rossoctl.io/protocol single-value label.
-    if not protocols:
-        legacy = labels.get("rossoctl.io/protocol")
-        if legacy:
-            protocols = [legacy]
-
-    return ResourceLabels(
-        protocol=protocols or None,
-        framework=labels.get("rossoctl.io/framework"),
-        type=labels.get("rossoctl.io/type"),
-    )
 
 
 @router.get(
@@ -1280,3897 +783,121 @@ async def delete_agent(
 # =============================================================================
 
 
-@router.get(
-    "/migration/migratable",
-    response_model=ListMigratableAgentsResponse,
-    summary="List agents that can be migrated from Agent CRD to Deployment",
-    tags=["migration"],
-    dependencies=[Depends(require_roles(ROLE_VIEWER))],
-)
-async def list_migratable_agents(
-    namespace: str = Query(default="default", description="Kubernetes namespace"),
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> ListMigratableAgentsResponse:
-    """
-    List all Agent CRDs in a namespace that can be migrated to Deployments.
-
-    Returns information about each agent including whether a Deployment
-    already exists (indicating migration is complete).
-    """
-    try:
-        # List legacy Agent CRDs
-        agent_crds = kube.list_custom_resources(
-            group=CRD_GROUP,
-            version=CRD_VERSION,
-            namespace=namespace,
-            plural=AGENTS_PLURAL,
-        )
-    except ApiException as e:
-        if e.status == 404:
-            # CRD not installed
-            return ListMigratableAgentsResponse(agents=[], total=0, already_migrated=0)
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-    # Get list of existing Deployments to check for already-migrated agents
-    try:
-        existing_deployments = kube.list_deployments(
-            namespace=namespace,
-            label_selector=f"{ROSSOCTL_TYPE_LABEL}={RESOURCE_TYPE_AGENT}",
-        )
-        existing_names = {d.get("metadata", {}).get("name") for d in existing_deployments}
-    except ApiException:
-        existing_names = set()
-
-    agents = []
-    already_migrated = 0
-
-    for agent in agent_crds:
-        metadata = agent.get("metadata", {})
-        name = metadata.get("name", "")
-        labels = metadata.get("labels", {})
-        has_deployment = name in existing_names
-
-        if has_deployment:
-            already_migrated += 1
-
-        # Get description from spec or annotations
-        spec = agent.get("spec", {})
-        description = spec.get("description") or metadata.get("annotations", {}).get(
-            ROSSOCTL_DESCRIPTION_ANNOTATION, ""
-        )
-
-        # Determine status
-        status = agent.get("status", {})
-        agent_status = "Unknown"
-        for cond in status.get("conditions") or []:
-            if cond.get("type") == "Ready":
-                agent_status = "Ready" if cond.get("status") == "True" else "Not Ready"
-                break
-
-        agents.append(
-            MigratableAgentInfo(
-                name=name,
-                namespace=namespace,
-                status=agent_status,
-                has_deployment=has_deployment,
-                labels=labels,
-                description=description,
-            )
-        )
-
-    return ListMigratableAgentsResponse(
-        agents=agents,
-        total=len(agents),
-        already_migrated=already_migrated,
-    )
-
-
-@router.post(
-    "/{namespace}/{name}/migrate",
-    response_model=MigrateAgentResponse,
-    summary="Migrate an Agent CRD to a Deployment",
-    tags=["migration"],
-    dependencies=[Depends(require_roles(ROLE_OPERATOR))],
-)
-async def migrate_agent(
-    namespace: str,
-    name: str,
-    request: MigrateAgentRequest = MigrateAgentRequest(),
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> MigrateAgentResponse:
-    """
-    Migrate an Agent CRD to a Deployment.
-
-    This endpoint:
-    1. Reads the existing Agent CRD specification
-    2. Creates a Deployment with the same pod template
-    3. Creates a Service for the Deployment
-    4. Optionally deletes the Agent CRD (if delete_old=True)
-
-    If a Deployment already exists with the same name, the migration will fail
-    unless the existing Deployment was created by rossoctl-operator (in which
-    case we just need to clean up the Agent CRD).
-    """
-    logger.info(f"Starting migration of Agent CRD '{name}' in namespace '{namespace}'")
-
-    deployment_created = False
-    service_created = False
-    agent_crd_deleted = False
-
-    # Step 1: Get the Agent CRD
-    try:
-        agent = kube.get_custom_resource(
-            group=CRD_GROUP,
-            version=CRD_VERSION,
-            namespace=namespace,
-            plural=AGENTS_PLURAL,
-            name=name,
-        )
-    except ApiException as e:
-        if e.status == 404:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Agent CRD '{name}' not found in namespace '{namespace}'",
-            )
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-    # Step 2: Check if Deployment already exists
-    deployment_exists = False
-    deployment_managed_by_operator = False
-    try:
-        existing_deployment = kube.get_deployment(namespace=namespace, name=name)
-        deployment_exists = True
-        # Check if it was created by rossoctl-operator
-        dep_labels = existing_deployment.get("metadata", {}).get("labels", {})
-        deployment_managed_by_operator = (
-            dep_labels.get(APP_KUBERNETES_IO_CREATED_BY) == ROSSOCTL_OPERATOR_LABEL_NAME
-            or dep_labels.get(APP_KUBERNETES_IO_MANAGED_BY) == ROSSOCTL_OPERATOR_LABEL_NAME
-        )
-        logger.info(
-            f"Deployment '{name}' already exists, managed_by_operator={deployment_managed_by_operator}"
-        )
-    except ApiException as e:
-        if e.status != 404:
-            raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-    # Step 3: Check if Service already exists
-    service_exists = False
-    try:
-        kube.get_service(namespace=namespace, name=name)
-        service_exists = True
-        logger.info(f"Service '{name}' already exists")
-    except ApiException as e:
-        if e.status != 404:
-            raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-    # Step 4: Build and create Deployment (if needed)
-    if deployment_exists:
-        if deployment_managed_by_operator:
-            # Deployment was created by operator, we just need to update labels
-            # to mark it as migrated (managed by rossoctl-ui now)
-            try:
-                patch = {
-                    "metadata": {
-                        "labels": {
-                            APP_KUBERNETES_IO_MANAGED_BY: ROSSOCTL_UI_CREATOR_LABEL,
-                        },
-                        "annotations": {
-                            MIGRATION_SOURCE_ANNOTATION: "agent-crd",
-                            MIGRATION_TIMESTAMP_ANNOTATION: datetime.now(timezone.utc).isoformat(),
-                        },
-                    }
-                }
-                kube.patch_deployment(namespace=namespace, name=name, body=patch)
-                logger.info(f"Patched Deployment '{name}' with migration annotations")
-            except ApiException as e:
-                logger.warning(f"Failed to patch Deployment '{name}': {e.reason}")
-        else:
-            raise HTTPException(
-                status_code=409,
-                detail=f"Deployment '{name}' already exists and was not created by rossoctl-operator. "
-                "Cannot migrate. Delete the existing Deployment first or use a different name.",
-            )
-    else:
-        # Create new Deployment from Agent CRD spec
-        deployment_manifest = _build_deployment_from_agent_crd(agent)
-        kube.ensure_service_account(namespace=namespace, name=name)
-        try:
-            kube.create_deployment(namespace=namespace, body=deployment_manifest)
-            deployment_created = True
-            logger.info(f"Created Deployment '{name}' from Agent CRD")
-        except ApiException as e:
-            raise HTTPException(
-                status_code=e.status,
-                detail=f"Failed to create Deployment: {e.reason}",
-            )
-
-    # Step 5: Build and create Service (if needed)
-    if not service_exists:
-        service_manifest = _build_service_from_agent_crd(agent)
-        try:
-            kube.create_service(namespace=namespace, body=service_manifest)
-            service_created = True
-            logger.info(f"Created Service '{name}' from Agent CRD")
-        except ApiException as e:
-            # If Deployment was created, try to clean up
-            if deployment_created:
-                try:
-                    kube.delete_deployment(namespace=namespace, name=name)
-                except Exception as cleanup_error:
-                    logger.warning(
-                        "Failed to clean up Deployment '%s' after Service creation error: %s",
-                        name,
-                        cleanup_error,
-                    )
-            raise HTTPException(
-                status_code=e.status,
-                detail=f"Failed to create Service: {e.reason}",
-            )
-
-    # Step 6: Delete the Agent CRD (if requested)
-    if request.delete_old:
-        try:
-            kube.delete_custom_resource(
-                group=CRD_GROUP,
-                version=CRD_VERSION,
-                namespace=namespace,
-                plural=AGENTS_PLURAL,
-                name=name,
-            )
-            agent_crd_deleted = True
-            logger.info(f"Deleted Agent CRD '{name}'")
-        except ApiException as e:
-            if e.status != 404:
-                logger.warning(f"Failed to delete Agent CRD '{name}': {e.reason}")
-
-    # Build response message
-    messages = []
-    if deployment_created:
-        messages.append("Deployment created")
-    elif deployment_exists and deployment_managed_by_operator:
-        messages.append("Deployment updated (was created by operator)")
-    if service_created:
-        messages.append("Service created")
-    elif service_exists:
-        messages.append("Service already exists")
-    if agent_crd_deleted:
-        messages.append("Agent CRD deleted")
-    elif request.delete_old:
-        messages.append("Agent CRD deletion requested but skipped")
-
-    return MigrateAgentResponse(
-        success=True,
-        migrated=True,
-        name=name,
-        namespace=namespace,
-        message="; ".join(messages) if messages else "Migration completed",
-        deployment_created=deployment_created,
-        service_created=service_created,
-        agent_crd_deleted=agent_crd_deleted,
-    )
-
-
-@router.post(
-    "/migration/migrate-all",
-    response_model=Dict[str, Any],
-    summary="Migrate all Agent CRDs in a namespace to Deployments",
-    tags=["migration"],
-    dependencies=[Depends(require_roles(ROLE_OPERATOR))],
-)
-async def migrate_all_agents(
-    namespace: str = Query(default="default", description="Kubernetes namespace"),
-    delete_old: bool = Query(default=False, description="Delete Agent CRDs after migration"),
-    dry_run: bool = Query(default=True, description="If True, only show what would be migrated"),
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> Dict[str, Any]:
-    """
-    Migrate all Agent CRDs in a namespace to Deployments.
-
-    Use dry_run=True (default) to see what would be migrated before actually performing
-    the migration. Set dry_run=False to execute the migration.
-    """
-    # First, get the list of migratable agents
-    migratable = await list_migratable_agents(namespace=namespace, kube=kube)
-
-    results = {
-        "namespace": namespace,
-        "dry_run": dry_run,
-        "delete_old": delete_old,
-        "total_agents": migratable.total,
-        "already_migrated": migratable.already_migrated,
-        "to_migrate": migratable.total - migratable.already_migrated,
-        "migrated": [],
-        "skipped": [],
-        "failed": [],
-    }
-
-    for agent_info in migratable.agents:
-        if agent_info.has_deployment:
-            results["skipped"].append(
-                {
-                    "name": agent_info.name,
-                    "reason": "Deployment already exists",
-                }
-            )
-            continue
-
-        if dry_run:
-            results["migrated"].append(
-                {
-                    "name": agent_info.name,
-                    "status": "would be migrated (dry-run)",
-                }
-            )
-        else:
-            try:
-                result = await migrate_agent(
-                    namespace=namespace,
-                    name=agent_info.name,
-                    request=MigrateAgentRequest(delete_old=delete_old),
-                    kube=kube,
-                )
-                results["migrated"].append(
-                    {
-                        "name": agent_info.name,
-                        "status": "migrated",
-                        "message": result.message,
-                    }
-                )
-            except HTTPException as e:
-                results["failed"].append(
-                    {
-                        "name": agent_info.name,
-                        "error": e.detail,
-                    }
-                )
-            except Exception as e:
-                results["failed"].append(
-                    {
-                        "name": agent_info.name,
-                        "error": str(e),
-                    }
-                )
-
-    return results
-
-
-def _build_deployment_from_agent_crd(agent: dict) -> dict:
-    """
-    Build a Kubernetes Deployment manifest from an Agent CRD.
-
-    Args:
-        agent: The Agent CRD resource dictionary.
-
-    Returns:
-        Deployment manifest dictionary.
-    """
-    metadata = agent.get("metadata", {})
-    spec = agent.get("spec", {})
-    name = metadata.get("name", "")
-    namespace = metadata.get("namespace", "default")
-
-    # Get labels from Agent CRD and update for Deployment
-    labels = metadata.get("labels", {}).copy()
-    labels[ROSSOCTL_WORKLOAD_TYPE_LABEL] = WORKLOAD_TYPE_DEPLOYMENT
-    labels[APP_KUBERNETES_IO_MANAGED_BY] = ROSSOCTL_UI_CREATOR_LABEL
-
-    # Get annotations
-    annotations = metadata.get("annotations", {}).copy()
-    annotations[MIGRATION_SOURCE_ANNOTATION] = "agent-crd"
-    annotations[MIGRATION_TIMESTAMP_ANNOTATION] = datetime.now(timezone.utc).isoformat()
-
-    # Description
-    description = spec.get("description", "")
-    if description:
-        annotations[ROSSOCTL_DESCRIPTION_ANNOTATION] = description
-
-    # Extract pod template from Agent CRD
-    pod_template_spec = spec.get("podTemplateSpec", {})
-    pod_spec = pod_template_spec.get("spec", {})
-
-    # If no pod template, try to build one from imageSource
-    if not pod_spec:
-        image_source = spec.get("imageSource", {})
-        image = image_source.get("image", "")
-        if not image:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Agent CRD '{name}' has no podTemplateSpec or imageSource.image",
-            )
-
-        pod_spec = {
-            "serviceAccountName": name,
-            "containers": [
-                {
-                    "name": "agent",
-                    "image": image,
-                    "imagePullPolicy": DEFAULT_IMAGE_POLICY,
-                    "resources": {
-                        "limits": DEFAULT_RESOURCE_LIMITS,
-                        "requests": DEFAULT_RESOURCE_REQUESTS,
-                    },
-                    "ports": [
-                        {
-                            "name": "http",
-                            "containerPort": DEFAULT_IN_CLUSTER_PORT,
-                            "protocol": "TCP",
-                        }
-                    ],
-                    "volumeMounts": [
-                        {"name": "cache", "mountPath": "/app/.cache"},
-                        {"name": "shared-data", "mountPath": "/shared"},
-                    ],
-                }
-            ],
-            "volumes": [
-                {"name": "cache", "emptyDir": {}},
-                {"name": "shared-data", "emptyDir": {}},
-            ],
-        }
-
-    # Ensure serviceAccountName is set so the webhook's SPIFFE identity
-    # derivation uses the workload name rather than the ReplicaSet hash.
-    pod_spec.setdefault("serviceAccountName", name)
-
-    # Build selector labels (type label is applied by the operator via AgentRuntime)
-    selector_labels = {
-        APP_KUBERNETES_IO_NAME: name,
-    }
-
-    # Build pod template labels (merge selector labels with other labels)
-    pod_labels = labels.copy()
-
-    # Get replicas
-    replicas = spec.get("replicas", 1)
-
-    return {
-        "apiVersion": "apps/v1",
-        "kind": "Deployment",
-        "metadata": {
-            "name": name,
-            "namespace": namespace,
-            "labels": labels,
-            "annotations": annotations,
-        },
-        "spec": {
-            "replicas": replicas,
-            "selector": {
-                "matchLabels": selector_labels,
-            },
-            "template": {
-                "metadata": {
-                    "labels": pod_labels,
-                },
-                "spec": pod_spec,
-            },
-        },
-    }
-
-
-def _build_service_from_agent_crd(agent: dict) -> dict:
-    """
-    Build a Kubernetes Service manifest from an Agent CRD.
-
-    Args:
-        agent: The Agent CRD resource dictionary.
-
-    Returns:
-        Service manifest dictionary.
-    """
-    metadata = agent.get("metadata", {})
-    spec = agent.get("spec", {})
-    name = metadata.get("name", "")
-    namespace = metadata.get("namespace", "default")
-
-    # Get labels
-    labels = metadata.get("labels", {}).copy()
-    labels[APP_KUBERNETES_IO_MANAGED_BY] = ROSSOCTL_UI_CREATOR_LABEL
-
-    # Build selector labels (type label is applied by the operator via AgentRuntime)
-    selector_labels = {
-        APP_KUBERNETES_IO_NAME: name,
-    }
-
-    # Get service ports from Agent CRD
-    service_ports_spec = spec.get("servicePorts", [])
-    if service_ports_spec:
-        service_ports = [
-            {
-                "name": sp.get("name", "http"),
-                "port": sp.get("port", DEFAULT_OFF_CLUSTER_PORT),
-                "targetPort": sp.get("targetPort", DEFAULT_IN_CLUSTER_PORT),
-                "protocol": sp.get("protocol", "TCP"),
-            }
-            for sp in service_ports_spec
-        ]
-    else:
-        service_ports = [
-            {
-                "name": "http",
-                "port": DEFAULT_OFF_CLUSTER_PORT,
-                "targetPort": DEFAULT_IN_CLUSTER_PORT,
-                "protocol": "TCP",
-            }
-        ]
-
-    return {
-        "apiVersion": "v1",
-        "kind": "Service",
-        "metadata": {
-            "name": name,
-            "namespace": namespace,
-            "labels": labels,
-        },
-        "spec": {
-            "type": "ClusterIP",
-            "selector": selector_labels,
-            "ports": service_ports,
-        },
-    }
-
-
-@router.get(
-    "/build-strategies",
-    response_model=ClusterBuildStrategiesResponse,
-    dependencies=[Depends(require_roles(ROLE_VIEWER))],
-)
-async def list_build_strategies(
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> ClusterBuildStrategiesResponse:
-    """List available ClusterBuildStrategies for Shipwright builds.
-
-    Returns the list of ClusterBuildStrategy resources available in the cluster.
-    """
-    try:
-        response = kube.list_cluster_custom_resources(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            plural=SHIPWRIGHT_CLUSTER_BUILD_STRATEGIES_PLURAL,
-        )
-
-        strategy_list = []
-        for strategy in response.get("items", []):
-            metadata = strategy.get("metadata", {})
-            spec = strategy.get("spec", {})
-            # Get description from annotations or spec
-            annotations = metadata.get("annotations", {})
-            description = annotations.get("description") or spec.get("description")
-
-            strategy_list.append(
-                ClusterBuildStrategyInfo(
-                    name=metadata.get("name", ""),
-                    description=description,
-                )
-            )
-
-        return ClusterBuildStrategiesResponse(strategies=strategy_list)
-
-    except ApiException as e:
-        logger.error(f"Failed to list ClusterBuildStrategies: {e}")
-        raise HTTPException(
-            status_code=e.status,
-            detail=f"Failed to list build strategies: {e.reason}",
-        )
-
-
-@router.get(
-    "/shipwright-builds",
-    response_model=ShipwrightBuildListResponse,
-    dependencies=[Depends(require_roles(ROLE_VIEWER))],
-)
-async def list_agent_shipwright_builds(
-    namespace: str = Query(
-        default="",
-        description="Kubernetes namespace (required unless all_namespaces=true)",
-    ),
-    all_namespaces: bool = Query(
-        default=False,
-        alias="allNamespaces",
-        description="If true, list builds in all rossoctl-enabled namespaces",
-    ),
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> ShipwrightBuildListResponse:
-    """List Shipwright Build resources for agents only (rossoctl.io/type=agent)."""
-    namespaces_to_scan: List[str] = []
-    if all_namespaces:
-        namespaces_to_scan = kube.list_enabled_namespaces()
-    else:
-        if not namespace or not namespace.strip():
-            raise HTTPException(
-                status_code=400,
-                detail="namespace query parameter is required (or use allNamespaces=true)",
-            )
-        namespaces_to_scan = [namespace.strip()]
-
-    try:
-        items = collect_rossoctl_shipwright_builds(
-            kube, namespaces_to_scan, RESOURCE_TYPE_AGENT, logger
-        )
-    except ApiException as e:
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-    return ShipwrightBuildListResponse(items=items)
-
-
-@router.get(
-    "/{namespace}/{name}/shipwright-build",
-    response_model=ShipwrightBuildStatusResponse,
-    dependencies=[Depends(require_roles(ROLE_VIEWER))],
-)
-async def get_shipwright_build_status(
-    namespace: str,
-    name: str,
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> ShipwrightBuildStatusResponse:
-    """Get the Shipwright Build status for an agent.
-
-    Returns the Build resource status including whether it's registered
-    and ready for BuildRuns.
-    """
-    try:
-        build = kube.get_custom_resource(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            namespace=namespace,
-            plural=SHIPWRIGHT_BUILDS_PLURAL,
-            name=name,
-        )
-
-        metadata = build.get("metadata", {})
-        status = build.get("status", {})
-
-        # Check if build is registered (strategy validated)
-        registered = status.get("registered", False)
-        reason = status.get("reason")
-        message = status.get("message")
-
-        return ShipwrightBuildStatusResponse(
-            name=metadata.get("name", name),
-            namespace=metadata.get("namespace", namespace),
-            registered=registered,
-            reason=reason,
-            message=message,
-        )
-
-    except ApiException as e:
-        if e.status == 404:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Shipwright Build '{name}' not found in namespace '{namespace}'",
-            )
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-
-@router.get(
-    "/{namespace}/{name}/shipwright-buildrun",
-    response_model=ShipwrightBuildRunStatusResponse,
-    dependencies=[Depends(require_roles(ROLE_VIEWER))],
-)
-async def get_shipwright_buildrun_status(
-    namespace: str,
-    name: str,
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> ShipwrightBuildRunStatusResponse:
-    """Get the latest Shipwright BuildRun status for an agent build.
-
-    Lists BuildRuns with label selector for the build name and returns
-    the most recent one's status.
-    """
-    try:
-        # List BuildRuns with label selector for this build
-        items = kube.list_custom_resources(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            namespace=namespace,
-            plural=SHIPWRIGHT_BUILDRUNS_PLURAL,
-            label_selector=f"rossoctl.io/build-name={name}",
-        )
-
-        if not items:
-            raise HTTPException(
-                status_code=404,
-                detail=f"No BuildRuns found for build '{name}' in namespace '{namespace}'",
-            )
-
-        # Sort by creation timestamp and get the most recent
-        items.sort(
-            key=lambda x: x.get("metadata", {}).get("creationTimestamp", ""),
-            reverse=True,
-        )
-        latest_buildrun = items[0]
-
-        metadata = latest_buildrun.get("metadata", {})
-        status = latest_buildrun.get("status", {})
-        spec = latest_buildrun.get("spec", {})
-
-        # Extract conditions
-        conditions = []
-        for cond in status.get("conditions") or []:
-            conditions.append(
-                BuildStatusCondition(
-                    type=cond.get("type", ""),
-                    status=cond.get("status", ""),
-                    reason=cond.get("reason"),
-                    message=cond.get("message"),
-                    lastTransitionTime=cond.get("lastTransitionTime"),
-                )
-            )
-
-        # Determine phase from conditions
-        phase = "Pending"
-        failure_message = None
-        for cond in conditions:
-            if cond.type == "Succeeded":
-                if cond.status == "True":
-                    phase = "Succeeded"
-                elif cond.status == "False":
-                    phase = "Failed"
-                    failure_message = cond.message
-                else:
-                    phase = "Running"
-                break
-
-        # Get output image info
-        output = status.get("output", {})
-        output_image = output.get("image")
-        output_digest = output.get("digest")
-
-        return ShipwrightBuildRunStatusResponse(
-            name=metadata.get("name", ""),
-            namespace=metadata.get("namespace", namespace),
-            buildName=spec.get("build", {}).get("name", name),
-            phase=phase,
-            startTime=status.get("startTime"),
-            completionTime=status.get("completionTime"),
-            outputImage=output_image,
-            outputDigest=output_digest,
-            failureMessage=failure_message,
-            conditions=conditions,
-        )
-
-    except ApiException as e:
-        if e.status == 404:
-            raise HTTPException(
-                status_code=404,
-                detail=f"BuildRun not found for build '{name}' in namespace '{namespace}'",
-            )
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-
-@router.post(
-    "/{namespace}/{name}/shipwright-buildrun", dependencies=[Depends(require_roles(ROLE_OPERATOR))]
-)
-async def trigger_shipwright_buildrun(
-    namespace: str,
-    name: str,
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> Dict[str, Any]:
-    """Trigger a new Shipwright BuildRun for an existing Build.
-
-    Creates a new BuildRun resource to start a build execution.
-    """
-    try:
-        # First verify the Build exists
-        build = kube.get_custom_resource(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            namespace=namespace,
-            plural=SHIPWRIGHT_BUILDS_PLURAL,
-            name=name,
-        )
-
-        # Get labels from the Build to propagate to BuildRun
-        build_labels = build.get("metadata", {}).get("labels", {})
-        buildrun_labels = {
-            k: v
-            for k, v in build_labels.items()
-            if k.startswith("rossoctl.io/") or k.startswith("app.kubernetes.io/")
-        }
-
-        # Create BuildRun manifest
-        buildrun_manifest = _build_agent_shipwright_buildrun_manifest(
-            build_name=name,
-            namespace=namespace,
-            labels=buildrun_labels,
-        )
-
-        # Create the BuildRun
-        created_buildrun = kube.create_custom_resource(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            namespace=namespace,
-            plural=SHIPWRIGHT_BUILDRUNS_PLURAL,
-            body=buildrun_manifest,
-        )
-
-        return {
-            "success": True,
-            "buildRunName": created_buildrun.get("metadata", {}).get("name"),
-            "namespace": namespace,
-            "buildName": name,
-            "message": "BuildRun created successfully",
-        }
-
-    except ApiException as e:
-        if e.status == 404:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Build '{name}' not found in namespace '{namespace}'",
-            )
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-
-@router.get(
-    "/{namespace}/{name}/shipwright-build-info",
-    response_model=AgentShipwrightBuildInfoResponse,
-    dependencies=[Depends(require_roles(ROLE_VIEWER))],
-)
-async def get_shipwright_build_info(
-    namespace: str,
-    name: str,
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> AgentShipwrightBuildInfoResponse:
-    """Get full Shipwright Build information including agent config and BuildRun status.
-
-    This endpoint provides all the information needed for the build progress page:
-    - Build configuration and status
-    - Latest BuildRun status
-    - Agent configuration stored in annotations
-    """
-    try:
-        # Get the Build resource
-        build = kube.get_custom_resource(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            namespace=namespace,
-            plural=SHIPWRIGHT_BUILDS_PLURAL,
-            name=name,
-        )
-
-        metadata = build.get("metadata", {})
-        spec = build.get("spec", {})
-        status = build.get("status", {})
-
-        # Extract build info
-        source = spec.get("source", {})
-        git_info = source.get("git", {})
-        strategy = spec.get("strategy", {})
-        output = spec.get("output", {})
-
-        # Parse agent config from annotations using shared utility
-        agent_config = extract_resource_config_from_build(build, ResourceType.AGENT)
-
-        # Build response with basic build info
-        response = AgentShipwrightBuildInfoResponse(
-            name=metadata.get("name", name),
-            namespace=metadata.get("namespace", namespace),
-            buildRegistered=status.get("registered", False),
-            buildReason=status.get("reason"),
-            buildMessage=status.get("message"),
-            outputImage=output.get("image", ""),
-            strategy=strategy.get("name", ""),
-            gitUrl=git_info.get("url", ""),
-            gitRevision=git_info.get("revision", ""),
-            contextDir=source.get("contextDir", ""),
-            agentConfig=agent_config,
-        )
-
-        # Try to get the latest BuildRun
-        try:
-            items = kube.list_custom_resources(
-                group=SHIPWRIGHT_CRD_GROUP,
-                version=SHIPWRIGHT_CRD_VERSION,
-                namespace=namespace,
-                plural=SHIPWRIGHT_BUILDRUNS_PLURAL,
-                label_selector=f"rossoctl.io/build-name={name}",
-            )
-
-            if items:
-                latest_buildrun = get_latest_buildrun(items)
-                if latest_buildrun:
-                    buildrun_info = extract_buildrun_info(latest_buildrun)
-
-                    response.hasBuildRun = True
-                    response.buildRunName = buildrun_info["name"]
-                    response.buildRunPhase = buildrun_info["phase"]
-                    response.buildRunStartTime = buildrun_info["startTime"]
-                    response.buildRunCompletionTime = buildrun_info["completionTime"]
-                    response.buildRunOutputImage = buildrun_info["outputImage"]
-                    response.buildRunOutputDigest = buildrun_info["outputDigest"]
-                    response.buildRunFailureMessage = buildrun_info["failureMessage"]
-
-        except ApiException as e:
-            # BuildRun not found is OK, just means no build has been triggered
-            if e.status != 404:
-                logger.warning(f"Failed to get BuildRun for build '{name}': {e}")
-
-        return response
-
-    except ApiException as e:
-        if e.status == 404:
-            raise HTTPException(
-                status_code=404,
-                detail=f"Shipwright Build '{name}' not found in namespace '{namespace}'",
-            )
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-
-def _get_authbridge_runtime_yaml() -> str:
-    """Read the authbridge runtime config from the Helm-managed ConfigMap.
-
-    The pipeline configuration is defined in values.yaml (authBridge.pipeline),
-    rendered by the Helm chart into the authbridge-runtime-config ConfigMap,
-    and mounted into the backend pod. This is the single source of truth.
-
-    Falls back to legacy in-memory construction when the file does not
-    exist (e.g. local development outside the cluster).
-    """
-    config_path = settings.authbridge_runtime_config_path
-    try:
-        with open(config_path, "r") as f:
-            content = f.read()
-        if content.strip():
-            return content
-    except FileNotFoundError:
-        logger.warning(
-            "AuthBridge runtime config not found at %s; "
-            "falling back to legacy in-memory construction",
-            config_path,
-        )
-    except OSError as e:
-        logger.warning(
-            "Failed to read AuthBridge runtime config from %s: %s; "
-            "falling back to legacy construction",
-            config_path,
-            e,
-        )
-
-    return _build_authbridge_runtime_yaml_fallback()
-
-
-def _build_authbridge_runtime_yaml(
-    keycloak_url: str,
-    realm: str,
-    issuer: str,
-    spire_enabled: bool = False,
-) -> str:
-    """Build authbridge runtime config YAML with explicit parameters.
-
-    This function is used by tests to generate authbridge configuration
-    with specific parameters, independent of the settings object.
-
-    Args:
-        keycloak_url: Internal Keycloak URL for JWKS and token exchange
-        realm: Keycloak realm name
-        issuer: Public issuer URL (must match JWT "iss" claim)
-        spire_enabled: Whether to enable SPIFFE/SPIRE identity
-
-    Returns:
-        YAML string containing the authbridge runtime configuration
-    """
-    identity_type = "spiffe" if spire_enabled else "client-secret"
-    identity: dict[str, str] = {"type": identity_type}
-    if identity_type == "spiffe":
-        identity["jwt_audience"] = issuer
-
-    config: dict[str, object] = {}
-    if spire_enabled:
-        config["spiffe"] = {}
-
-    config["pipeline"] = {
-        "inbound": {
-            "plugins": [
-                {
-                    "name": "jwt-validation",
-                    "config": {
-                        "issuer": issuer,
-                        "keycloak_url": keycloak_url,
-                        "keycloak_realm": realm,
-                    },
-                }
-            ]
-        },
-        "outbound": {
-            "plugins": [
-                {
-                    "name": "token-exchange",
-                    "config": {
-                        "keycloak_url": keycloak_url,
-                        "keycloak_realm": realm,
-                        "default_policy": "passthrough",
-                        "identity": identity,
-                    },
-                }
-            ]
-        },
-    }
-
-    return yaml.dump(config, default_flow_style=False)
-
-
-def _build_authbridge_runtime_yaml_fallback() -> str:
-    """Legacy fallback: construct authbridge runtime config in-memory.
-
-    Used when the mounted ConfigMap is unavailable (local dev, tests).
-    The canonical source of truth is values.yaml authBridge.pipeline,
-    rendered by the Helm chart and mounted at the path specified by
-    settings.authbridge_runtime_config_path.
-    """
-    keycloak_url = settings.keycloak_url or DEFAULT_KEYCLOAK_INTERNAL_URL
-    realm = settings.effective_keycloak_realm or DEFAULT_KEYCLOAK_REALM
-    issuer = f"{settings.effective_keycloak_url}/realms/{realm}"
-    spire_enabled = False
-
-    identity_type = "spiffe" if spire_enabled else "client-secret"
-    identity: dict[str, str] = {"type": identity_type}
-    if identity_type == "spiffe":
-        identity["jwt_audience"] = issuer
-    config: dict[str, object] = {}
-    if spire_enabled:
-        config["spiffe"] = {}
-    config["pipeline"] = {
-        "inbound": {
-            "plugins": [
-                {
-                    "name": "jwt-validation",
-                    "config": {
-                        "issuer": issuer,
-                        "keycloak_url": keycloak_url,
-                        "keycloak_realm": realm,
-                    },
-                }
-            ]
-        },
-        "outbound": {
-            "plugins": [
-                {
-                    "name": "token-exchange",
-                    "config": {
-                        "keycloak_url": keycloak_url,
-                        "keycloak_realm": realm,
-                        "default_policy": "passthrough",
-                        "identity": identity,
-                    },
-                }
-            ]
-        },
-    }
-
-    return yaml.dump(config, default_flow_style=False)
-
-
-def _ensure_authbridge_configmaps(
-    kube: KubernetesService,
-    namespace: str,
-    spire_enabled: bool = False,
-) -> None:
-    """Ensure the 4 ConfigMaps required by AuthBridge sidecars exist.
-
-    Creates each ConfigMap only if it does not already exist, so user
-    customizations (e.g. pointing at a different Keycloak server) are
-    preserved on subsequent agent deploys.
-
-    ConfigMaps created:
-      - authbridge-config: flat key-value Keycloak URLs for client-registration
-      - authbridge-runtime-config: YAML config for the unified authbridge binary
-      - envoy-config: Envoy proxy listeners and ext-proc integration
-      - spiffe-helper-config: SPIFFE workload API socket paths and SVID output
-
-    For Helm-managed namespaces, the Helm chart creates equivalent
-    ConfigMaps at install time (see agent-namespaces.yaml).
-    """
-    keycloak_url = settings.keycloak_url or DEFAULT_KEYCLOAK_INTERNAL_URL
-    realm = settings.effective_keycloak_realm or DEFAULT_KEYCLOAK_REALM
-    # ISSUER must use the public/external URL because it must match the
-    # "iss" claim in JWT tokens issued by Keycloak (split-horizon DNS).
-    issuer = f"{settings.effective_keycloak_url}/realms/{realm}"
-
-    # 1. authbridge-config (flat key-value for client-registration and legacy go-processor)
-    kube.ensure_configmap(
-        namespace=namespace,
-        name="authbridge-config",
-        data={
-            "KEYCLOAK_URL": keycloak_url,
-            "KEYCLOAK_REALM": realm,
-            "ISSUER": issuer,
-            "SPIRE_ENABLED": "true" if spire_enabled else "false",
-        },
-    )
-
-    # 2. authbridge-runtime-config (YAML config for the unified authbridge binary)
-    # The operator reads this at admission time and creates a per-agent ConfigMap
-    # with mode and listener addresses merged in.
-    # Source of truth: values.yaml authBridge.pipeline → Helm renders ConfigMap →
-    # mounted into this pod → _get_authbridge_runtime_yaml() reads it.
-    kube.ensure_configmap(
-        namespace=namespace,
-        name="authbridge-runtime-config",
-        data={"config.yaml": _get_authbridge_runtime_yaml()},
-    )
-
-    # 3. envoy-config
-    kube.ensure_configmap(
-        namespace=namespace,
-        name="envoy-config",
-        data={"envoy.yaml": DEFAULT_ENVOY_YAML},
-    )
-
-    # 4. spiffe-helper-config
-    kube.ensure_configmap(
-        namespace=namespace,
-        name="spiffe-helper-config",
-        data={"helper.conf": DEFAULT_SPIFFE_HELPER_CONF},
-    )
-
-    logger.info(f"Ensured AuthBridge ConfigMaps in namespace '{namespace}'")
-
-
-def _ensure_authproxy_routes(
-    kube: KubernetesService,
-    namespace: str,
-    routes: List["OutboundRoute"],
-) -> None:
-    """Create or update the authproxy-routes ConfigMap with outbound token exchange rules."""
-    import yaml as _yaml
-
-    # AuthProxy go-processor expects a YAML list at file root (static.go), not {"routes": [...]}.
-    routes_list = [r.model_dump() for r in routes]
-    kube.upsert_configmap(
-        namespace=namespace,
-        name="authproxy-routes",
-        data={"routes.yaml": _yaml.dump(routes_list, default_flow_style=False)},
-    )
-    logger.info(
-        "Upserted authproxy-routes ConfigMap in namespace '%s' with %d route(s)",
-        namespace,
-        len(routes),
-    )
-
-
-def _ensure_authbridge_scc_rolebinding(
-    kube: KubernetesService,
-    namespace: str,
-) -> None:
-    """On OpenShift, ensure the AuthBridge SCC RoleBinding exists.
-
-    AuthBridge sidecars need NET_ADMIN/NET_RAW capabilities, RunAsAny UIDs,
-    and CSI volumes that OpenShift's default restricted-v2 SCC blocks.
-    The Helm chart creates the ``rossoctl-authbridge`` SCC and its ClusterRole;
-    this function creates the per-namespace RoleBinding that grants it to all
-    service accounts in the namespace.
-
-    On non-OpenShift clusters this is a no-op.  If the ClusterRole doesn't
-    exist (SCC not installed), a warning is logged and the function returns
-    without error — the agent will still be created, but pods may fail with
-    SCC errors until the SCC is installed.
-    """
-    if detect_platform(kube) != "openshift":
-        return
-
-    cluster_role_name = "system:openshift:scc:rossoctl-authbridge"
-
-    # Verify the ClusterRole exists (implies the SCC was installed)
-    try:
-        kube.rbac_api.read_cluster_role(name=cluster_role_name)
-    except ApiException as e:
-        if e.status == 404:
-            logger.warning(
-                "ClusterRole '%s' not found. "
-                "The rossoctl-authbridge SCC may not be installed. "
-                "Agent pods may fail with SCC errors. "
-                "Install via: helm upgrade rossoctl charts/rossoctl --set openshift=true",
-                cluster_role_name,
-            )
-            return
-        raise
-
-    kube.ensure_rolebinding(
-        namespace=namespace,
-        name="agent-authbridge-scc",
-        cluster_role_name=cluster_role_name,
-        subjects=[
-            kubernetes.client.RbacV1Subject(
-                kind="Group",
-                api_group="rbac.authorization.k8s.io",
-                name=f"system:serviceaccounts:{namespace}",
-            ),
-        ],
-    )
-
-
-def _load_agent_skill_summaries(
-    kube: KubernetesService,
-    namespace: str,
-    skill_names: List[str],
-) -> List[Dict[str, Any]]:
-    """Load skill metadata from ConfigMaps referenced by an agent.
-
-    The agent annotation stores user-facing skill names. For each skill, look up
-    the matching skill ConfigMap by either display-name annotation or resource name.
-    Missing skills are ignored so agent creation does not fail when a referenced
-    skill is deleted later.
-    """
-    if not skill_names:
-        return []
-
-    try:
-        cms = kube.core_api.list_namespaced_config_map(
-            namespace=namespace,
-            label_selector=f"{SKILL_TYPE_LABEL}={SKILL_TYPE_VALUE}",
-        )
-    except ApiException as exc:
-        # Sanitize namespace to prevent log injection
-        safe_namespace = namespace.replace("\n", "\\n").replace("\r", "\\r")
-        logger.warning(
-            "Failed to list skills for agent card generation: %s",
-            exc,
-            extra={"namespace": safe_namespace},
-        )
-        return []
-
-    requested = {skill_name.strip() for skill_name in skill_names if skill_name.strip()}
-    if not requested:
-        return []
-
-    summaries: List[Dict[str, Any]] = []
-    for cm in cms.items:
-        annotations = cm.metadata.annotations or {}
-        display_name = annotations.get(SKILL_DISPLAY_NAME_ANNOTATION) or cm.metadata.name
-        if display_name not in requested and cm.metadata.name not in requested:
-            continue
-
-        summaries.append(
-            {
-                "id": cm.metadata.name,
-                "name": display_name,
-                "description": annotations.get(SKILL_DESCRIPTION_ANNOTATION, ""),
-                "examples": [],
-            }
-        )
-
-    summaries.sort(key=lambda skill: skill["name"].lower())
-    return summaries
-
-
-def _ensure_card_unsigned_configmap(
-    kube: KubernetesService,
-    name: str,
-    namespace: str,
-    service_port: int = DEFAULT_IN_CLUSTER_PORT,
-    description: Optional[str] = None,
-    skill_names: Optional[List[str]] = None,
-) -> None:
-    """Create the <agent>-card-unsigned ConfigMap if it does not exist.
-
-    The Rossoctl operator webhook checks for this ConfigMap when a
-    Deployment is admitted.  If it exists, the webhook injects a
-    ``sign-agentcard`` init container that signs the agent card with
-    the workload's SPIRE SVID.  The ConfigMap must therefore be
-    created **before** the Deployment.
-    """
-    agent_url = f"http://{name}.{namespace}.svc.cluster.local:{service_port}"
-    skills = _load_agent_skill_summaries(kube, namespace, skill_names or [])
-    agent_card = json.dumps(
-        {
-            "name": name,
-            "description": description,
-            "url": agent_url,
-            "version": "1.0.0",
-            "capabilities": {},
-            "defaultInputModes": ["application/json"],
-            "defaultOutputModes": ["text/plain"],
-            "skills": skills,
-        },
-        indent=2,
-    )
-    kube.ensure_configmap(
-        namespace=namespace,
-        name=f"{name}-card-unsigned",
-        data={"agent.json": agent_card},
-    )
-
-
-def _build_agent_shipwright_build_manifest(
-    request: CreateAgentRequest, clone_secret_name: Optional[str] = None
-) -> dict:
-    """
-    Build a Shipwright Build CRD manifest for building an agent from source.
-
-    This is a wrapper around the shared build_shipwright_build_manifest function
-    that converts CreateAgentRequest to the shared function's parameters.
-    """
-    # Determine registry URL
-    registry_url = request.registryUrl or DEFAULT_INTERNAL_REGISTRY
-
-    # Build source config
-    source_config = BuildSourceConfig(
-        gitUrl=request.gitUrl,
-        gitRevision=request.gitBranch,
-        contextDir=request.gitPath or ".",
-        gitSecretName=clone_secret_name,
-    )
-
-    # Build output config
-    output_config = BuildOutputConfig(
-        registry=registry_url,
-        imageName=request.name,
-        imageTag=request.imageTag,
-        pushSecretName=request.registrySecret,
-    )
-
-    # Build resource configuration to store in annotation
-    resource_config: Dict[str, Any] = {
-        "protocol": request.protocol,
-        "framework": request.framework,
-        "createHttpRoute": request.createHttpRoute,
-        "registrySecret": request.registrySecret,
-        "workloadType": request.workloadType,  # Store workload type for finalization
-        "authBridgeEnabled": request.authBridgeEnabled,
-        "spireEnabled": request.spireEnabled,
-        "authBridgeMode": request.authBridgeMode,
-        "gitPath": request.gitPath,
-    }
-    if request.outboundRoutes:
-        resource_config["outboundRoutes"] = [r.model_dump() for r in request.outboundRoutes]
-    if request.outboundPortsExclude:
-        resource_config["outboundPortsExclude"] = request.outboundPortsExclude
-    if request.inboundPortsExclude:
-        resource_config["inboundPortsExclude"] = request.inboundPortsExclude
-    if request.defaultOutboundPolicy:
-        resource_config["defaultOutboundPolicy"] = request.defaultOutboundPolicy
-    if request.mtlsMode:
-        resource_config["mtlsMode"] = request.mtlsMode
-    if request.tlsBridgeEnabled:
-        resource_config["tlsBridgeEnabled"] = True
-    if request.persistentStorage:
-        resource_config["persistentStorage"] = request.persistentStorage.model_dump()
-    if request.k8sResourceLimits is not None:
-        resource_config["k8sResourceLimits"] = request.k8sResourceLimits
-    if request.k8sResourceRequests is not None:
-        resource_config["k8sResourceRequests"] = request.k8sResourceRequests
-    if request.contexts:
-        resource_config["contexts"] = [attachment.model_dump() for attachment in request.contexts]
-    # Add env vars if present
-    if request.envVars:
-        resource_config["envVars"] = [ev.model_dump(exclude_none=True) for ev in request.envVars]
-    if request.mcpToolName:
-        resource_config["mcpToolName"] = request.mcpToolName
-    if request.llmPreset:
-        resource_config["llmPreset"] = request.llmPreset
-    if request.llmModel:
-        resource_config["llmModel"] = request.llmModel
-    if request.skills:
-        resource_config["skills"] = request.skills
-    # Add service ports if present
-    if request.servicePorts:
-        resource_config["servicePorts"] = [sp.model_dump() for sp in request.servicePorts]
-
-    return build_shipwright_build_manifest(
-        name=request.name,
-        namespace=request.namespace,
-        resource_type=ResourceType.AGENT,
-        source_config=source_config,
-        output_config=output_config,
-        build_config=request.shipwrightConfig,
-        resource_config=resource_config,
-        protocol=request.protocol,
-        framework=request.framework,
-    )
-
-
-def _build_agent_shipwright_buildrun_manifest(
-    build_name: str, namespace: str, labels: Optional[Dict[str, str]] = None
-) -> dict:
-    """
-    Build a Shipwright BuildRun CRD manifest to trigger an agent build.
-
-    This is a wrapper around the shared build_shipwright_buildrun_manifest function.
-    """
-    return build_shipwright_buildrun_manifest(
-        build_name=build_name,
-        namespace=namespace,
-        resource_type=ResourceType.AGENT,
-        labels=labels,
-    )
-
-
-# -----------------------------------------------------------------------------
-# Workload Manifest Builders (Phase 1 - Migration to Standard K8s Workloads)
-# -----------------------------------------------------------------------------
-
-
-def _get_linked_skill_mounts(
-    request: "CreateAgentRequest",
-    skills_override: Optional[List[str]] = None,
-) -> tuple[List[Dict[str, Any]], List[Dict[str, Any]], Optional[str]]:
-    """Build volume and mount definitions for linked skill ConfigMaps (local only)."""
-    skills = skills_override if skills_override is not None else (request.skills or [])
-    if not skills:
-        return [], [], None
-
-    volumes: List[Dict[str, Any]] = []
-    volume_mounts: List[Dict[str, Any]] = []
-    skill_paths: List[str] = []
-
-    for index, skill_name in enumerate(skills):
-        if not skill_name:
-            continue
-        cm_name = _sanitize_k8s_name(skill_name)
-        volume_name = f"skill-{index}"
-        mount_path = f"{AGENT_SKILLS_MOUNT_ROOT}/{cm_name}"
-        volumes.append(
-            {
-                "name": volume_name,
-                "configMap": {
-                    "name": cm_name,
-                },
-            }
-        )
-        volume_mounts.append(
-            {
-                "name": volume_name,
-                "mountPath": mount_path,
-                "readOnly": True,
-            }
-        )
-        skill_paths.append(mount_path)
-
-    if not skill_paths:
-        return [], [], None
-
-    return volumes, volume_mounts, ",".join(skill_paths)
-
-
-def _is_skill_external(kube: "KubernetesService", namespace: str, skill_name: str) -> bool:
-    """Return True if the named skill ConfigMap is an external registry reference."""
-    try:
-        cm = kube.core_api.read_namespaced_config_map(
-            name=_sanitize_k8s_name(skill_name), namespace=namespace
-        )
-        labels = cm.metadata.labels or {}
-        return labels.get(SKILL_SOURCE_LABEL) == SKILL_SOURCE_EXTERNAL
-    except ApiException:
-        return False
-
-
-_SKILLBERRY_SH = """\
-#!/bin/sh
-set -e
-
-apk add -q --no-cache curl unzip 2>/dev/null || true
-
-URL="${REGISTRY_URL}/skills/${SKILL_NAME}/export-anthropic"
-
-echo "Fetching ${SKILL_NAME} from ${URL}"
-
-RETRIES=3
-DELAY=2
-for i in $(seq 1 $RETRIES); do
-    if curl -fsSL --max-filesize 52428800 -o /tmp/skill.zip "${URL}"; then
-        break
-    fi
-    if [ "$i" -eq "$RETRIES" ]; then
-        echo "FATAL: fetch failed after ${RETRIES} attempts"
-        exit 1
-    fi
-    echo "Attempt ${i} failed; retrying in ${DELAY}s..."
-    sleep $DELAY
-done
-
-mkdir -p "${TARGET_DIR}" /tmp/skill-extract
-unzip -q /tmp/skill.zip -d /tmp/skill-extract/
-SKILL_DIR=$(ls /tmp/skill-extract/ | head -1)
-cp -r "/tmp/skill-extract/${SKILL_DIR}/." "${TARGET_DIR}/"
-echo "OK: ${SKILL_NAME} -> ${TARGET_DIR}"
-"""
-
-_GENERIC_SH = """\
-#!/bin/sh
-set -e
-
-apk add -q --no-cache curl 2>/dev/null || true
-echo "Fetching skill from ${REGISTRY_URL}"
-
-RETRIES=3
-DELAY=2
-for i in $(seq 1 $RETRIES); do
-    if curl -fsSL --max-filesize 52428800 -o /tmp/skill.tar.gz "${REGISTRY_URL}"; then
-        break
-    fi
-    if [ "$i" -eq "$RETRIES" ]; then
-        echo "FATAL: fetch failed after ${RETRIES} attempts"
-        exit 1
-    fi
-    echo "Attempt ${i} failed; retrying in ${DELAY}s..."
-    sleep $DELAY
-done
-
-mkdir -p "${TARGET_DIR}"
-tar -xzf /tmp/skill.tar.gz -C "${TARGET_DIR}"
-echo "OK: ${REGISTRY_URL} -> ${TARGET_DIR}"
-"""
-
-
-def _build_fetcher_scripts_data() -> Dict[str, str]:
-    """Return ConfigMap data dict containing per-registry-type fetch scripts."""
-    return {"skillberry.sh": _SKILLBERRY_SH, "generic.sh": _GENERIC_SH}
-
-
-def _ensure_fetcher_scripts_cm(kube: "KubernetesService", namespace: str) -> None:
-    """Create or replace the rossoctl-skill-fetcher-scripts ConfigMap in namespace."""
-    import kubernetes.client as k8s_client
-
-    body = k8s_client.V1ConfigMap(
-        metadata=k8s_client.V1ObjectMeta(
-            name=SKILL_FETCHER_SCRIPTS_CM,
-            namespace=namespace,
-            labels={"app.kubernetes.io/managed-by": "rossoctl"},
-        ),
-        data=_build_fetcher_scripts_data(),
-    )
-    try:
-        kube.core_api.read_namespaced_config_map(name=SKILL_FETCHER_SCRIPTS_CM, namespace=namespace)
-        kube.core_api.replace_namespaced_config_map(
-            name=SKILL_FETCHER_SCRIPTS_CM, namespace=namespace, body=body
-        )
-    except ApiException as e:
-        if e.status == 404:
-            kube.core_api.create_namespaced_config_map(namespace=namespace, body=body)
-        else:
-            raise
-
-
-def _get_external_skill_data(
-    kube: "KubernetesService",
-    namespace: str,
-    all_skills: List[str],
-) -> tuple[List[Dict[str, Any]], List[Dict[str, Any]], List[Dict[str, Any]], List[str]]:
-    """Build init containers, volumes, and mounts for external-registry skills.
-
-    Returns (init_containers, volumes, main_volume_mounts, skill_paths).
-    One fetcher-scripts-vol is shared across all init containers in the pod.
-    """
-    init_containers: List[Dict[str, Any]] = []
-    volumes: List[Dict[str, Any]] = []
-    main_mounts: List[Dict[str, Any]] = []
-    skill_paths: List[str] = []
-    fetcher_vol_added = False
-
-    for index, skill_name in enumerate(all_skills):
-        if not skill_name:
-            continue
-        if not _is_skill_external(kube, namespace, skill_name):
-            continue
-
-        try:
-            cm = kube.core_api.read_namespaced_config_map(
-                name=_sanitize_k8s_name(skill_name), namespace=namespace
-            )
-        except ApiException:
-            continue
-
-        cm_labels = cm.metadata.labels or {}
-        cm_annotations = cm.metadata.annotations or {}
-        registry_type = cm_labels.get(SKILL_REGISTRY_TYPE_LABEL, "generic")
-        registry_url = cm_annotations.get(SKILL_REGISTRY_URL_ANNOTATION, "")
-        registry_skill_name = cm_annotations.get(SKILL_REGISTRY_SKILL_NAME_ANNOTATION, skill_name)
-        registry_skill_version = cm_annotations.get(
-            SKILL_REGISTRY_SKILL_VERSION_ANNOTATION, "latest"
-        )
-
-        cm_name = _sanitize_k8s_name(skill_name)
-        emptydir_vol_name = f"skill-ext-{index}"
-        mount_path = f"{AGENT_SKILLS_MOUNT_ROOT}/{cm_name}"
-
-        if not fetcher_vol_added:
-            volumes.append(
-                {
-                    "name": "fetcher-scripts-vol",
-                    "configMap": {"name": SKILL_FETCHER_SCRIPTS_CM},
-                }
-            )
-            fetcher_vol_added = True
-
-        volumes.append({"name": emptydir_vol_name, "emptyDir": {}})
-
-        init_containers.append(
-            {
-                "name": f"fetch-skill-{index}",
-                "image": SKILL_FETCHER_IMAGE,
-                "command": [
-                    "/bin/sh",
-                    "-c",
-                    (
-                        "SCRIPT=/fetcher-scripts/${REGISTRY_TYPE}.sh; "
-                        '[ -f "$SCRIPT" ] || SCRIPT=/fetcher-scripts/generic.sh; '
-                        '/bin/sh "$SCRIPT"'
-                    ),
-                ],
-                "env": [
-                    {"name": "REGISTRY_TYPE", "value": registry_type},
-                    {"name": "REGISTRY_URL", "value": registry_url},
-                    {"name": "SKILL_NAME", "value": registry_skill_name},
-                    {"name": "SKILL_VERSION", "value": registry_skill_version},
-                    {"name": "TARGET_DIR", "value": mount_path},
-                ],
-                "resources": {
-                    "requests": {"memory": "32Mi", "cpu": "50m"},
-                    "limits": {"memory": "128Mi", "cpu": "200m"},
-                },
-                "volumeMounts": [
-                    {"name": emptydir_vol_name, "mountPath": mount_path},
-                    {
-                        "name": "fetcher-scripts-vol",
-                        "mountPath": "/fetcher-scripts",
-                        "readOnly": True,
-                    },
-                ],
-            }
-        )
-
-        main_mounts.append(
-            {
-                "name": emptydir_vol_name,
-                "mountPath": mount_path,
-                "readOnly": True,
-            }
-        )
-        skill_paths.append(mount_path)
-
-    return init_containers, volumes, main_mounts, skill_paths
-
-
-def _build_env_vars(
-    request: "CreateAgentRequest",
-    local_skills: Optional[List[str]] = None,
-    ext_skill_paths: Optional[List[str]] = None,
-) -> List[dict]:
-    """
-    Build environment variables list with support for valueFrom references.
-
-    Args:
-        request: The agent creation request containing envVars.
-        local_skills: Optional override list of local skill names.
-        ext_skill_paths: Optional list of external skill mount paths.
-
-    Returns:
-        List of environment variable dictionaries.
-    """
-    env_vars = list(DEFAULT_ENV_VARS)
-    service_port = (
-        request.servicePorts[0].port if request.servicePorts else DEFAULT_OFF_CLUSTER_PORT
-    )
-    env_vars.append(
-        {
-            "name": AGENT_ENDPOINT,
-            "value": get_agent_url(request.name, request.namespace, service_port),
-        }
-    )
-
-    _, _, local_folders = _get_linked_skill_mounts(request, skills_override=local_skills)
-    all_paths = ([local_folders] if local_folders else []) + (ext_skill_paths or [])
-    if all_paths:
-        env_vars.append({"name": "SKILL_FOLDERS", "value": ",".join(all_paths)})
-
-    if request.envVars:
-        for ev in request.envVars:
-            if ev.value is not None:
-                # Direct value
-                env_vars.append({"name": ev.name, "value": ev.value})
-            elif ev.valueFrom is not None:
-                # Reference to Secret or ConfigMap
-                env_entry: Dict[str, Any] = {"name": ev.name, "valueFrom": {}}
-
-                if ev.valueFrom.secretKeyRef:
-                    env_entry["valueFrom"]["secretKeyRef"] = {
-                        "name": ev.valueFrom.secretKeyRef.name,
-                        "key": ev.valueFrom.secretKeyRef.key,
-                    }
-                elif ev.valueFrom.configMapKeyRef:
-                    env_entry["valueFrom"]["configMapKeyRef"] = {
-                        "name": ev.valueFrom.configMapKeyRef.name,
-                        "key": ev.valueFrom.configMapKeyRef.key,
-                    }
-
-                env_vars.append(env_entry)
-
-    # Deduplicate environment variables, keeping the last occurrence.
-    # Precedence (last wins): DEFAULT_ENV_VARS < AGENT_ENDPOINT/SKILL_FOLDERS < user envVars.
-    # User overrides of AGENT_ENDPOINT or SKILL_FOLDERS are intentional (advanced use).
-    seen = {}
-    for env in env_vars:
-        seen[env["name"]] = env
-    return list(seen.values())
-
-
-def _build_common_labels(
-    request: "CreateAgentRequest",
-    workload_type: str = WORKLOAD_TYPE_DEPLOYMENT,
-) -> Dict[str, str]:
-    """
-    Build common labels for agent workloads.
-
-    Common labels for agent workloads. The rossoctl.io/type label is applied
-    by the rossoctl-operator via AgentRuntime reconciliation, not here.
-
-    Args:
-        request: The agent creation request.
-        workload_type: The type of workload (deployment, statefulset, job).
-
-    Returns:
-        Dictionary of labels.
-    """
-    labels = {
-        APP_KUBERNETES_IO_NAME: request.name,
-        ROSSOCTL_FRAMEWORK_LABEL: request.framework,
-        ROSSOCTL_WORKLOAD_TYPE_LABEL: workload_type,
-        APP_KUBERNETES_IO_MANAGED_BY: ROSSOCTL_UI_CREATOR_LABEL,
-        APP_KUBERNETES_IO_COMPONENT: RESOURCE_TYPE_AGENT,
-        # AuthBridge sidecar injection control
-        ROSSOCTL_INJECT_LABEL: "enabled" if request.authBridgeEnabled else "disabled",
-    }
-    # Protocol label(s) using new prefix format
-    if request.protocol:
-        labels[f"{PROTOCOL_LABEL_PREFIX}{request.protocol}"] = ""
-    # SPIRE identity label — the operator's webhook reads this to set
-    # SPIRE_ENABLED=true on the combined sidecar's spiffe-helper.
-    if request.spireEnabled:
-        labels[ROSSOCTL_SPIRE_LABEL] = ROSSOCTL_SPIRE_ENABLED_VALUE
-    return labels
-
-
-def _build_common_annotations(request: "CreateAgentRequest") -> Dict[str, str]:
-    """Build pod template annotations for port exclusions and other webhook directives."""
-    annotations: Dict[str, str] = {}
-    if request.outboundPortsExclude:
-        annotations[ROSSOCTL_OUTBOUND_PORTS_EXCLUDE] = request.outboundPortsExclude
-    if request.inboundPortsExclude:
-        annotations[ROSSOCTL_INBOUND_PORTS_EXCLUDE] = request.inboundPortsExclude
-    return annotations
-
-
-def build_container_resources(
-    limits: Optional[Dict[str, str]] = None,
-    requests: Optional[Dict[str, str]] = None,
-) -> Dict[str, Dict[str, str]]:
-    """Build a container "resources" block from optional per-workload overrides.
-
-    None means "not specified" and selects the platform default. An empty dict is
-    honored as-is, because {} is how Kubernetes spells "no limits" -- it yields an
-    unbounded container rather than one capped at DEFAULT_RESOURCE_LIMITS. Testing
-    truthiness instead (`limits or DEFAULT_RESOURCE_LIMITS`) would collapse those
-    two cases and silently cap a workload the caller asked to leave uncapped.
-    """
-    return {
-        "limits": DEFAULT_RESOURCE_LIMITS if limits is None else limits,
-        "requests": DEFAULT_RESOURCE_REQUESTS if requests is None else requests,
-    }
-
-
-def _build_selector_labels(request: "CreateAgentRequest") -> Dict[str, str]:
-    """
-    Build selector labels for matching pods to workloads and services.
-
-    Args:
-        request: The agent creation request.
-
-    Returns:
-        Dictionary of selector labels.
-    """
-    return {
-        APP_KUBERNETES_IO_NAME: request.name,
-    }
-
-
-def _agentruntime_supported_workload(workload_type: str) -> bool:
-    """Whether a workload type gets an AgentRuntime CR (per-agent AuthBridge
-    config). Sandbox, deployment, and statefulset are supported; Job is not —
-    a run-to-completion Job doesn't fit the attach / config-rollout model.
-    Single source of truth for the two _ensure_agentruntime call sites
-    (create_agent and finalize_shipwright_build)."""
-    return workload_type not in (WORKLOAD_TYPE_JOB,)
-
-
-def _build_agentruntime_manifest(
-    name: str,
-    namespace: str,
-    workload_type: str,
-    agent_type: str = RESOURCE_TYPE_AGENT,
-    auth_bridge_mode: Optional[str] = None,
-    mtls_mode: Optional[str] = None,
-    tls_bridge_enabled: bool = False,
-) -> dict:
-    """Build an AgentRuntime CR manifest for the given workload."""
-    kind_map = {
-        WORKLOAD_TYPE_DEPLOYMENT: "Deployment",
-        WORKLOAD_TYPE_STATEFULSET: "StatefulSet",
-        WORKLOAD_TYPE_SANDBOX: "Sandbox",
-    }
-    # Sandbox is an agents.x-k8s.io CR, not apps/v1 — emit the right targetRef
-    # apiVersion per kind so the operator's resolveTargetRef finds the workload
-    # (a wrong apps/v1 ref for a Sandbox would dangle and never reconcile).
-    apiversion_map = {
-        WORKLOAD_TYPE_SANDBOX: f"{AGENT_SANDBOX_CRD_GROUP}/{AGENT_SANDBOX_CRD_VERSION}",
-    }
-    spec: dict = {
-        "type": agent_type,
-        "targetRef": {
-            "apiVersion": apiversion_map.get(workload_type, "apps/v1"),
-            "kind": kind_map.get(workload_type, "Deployment"),
-            "name": name,
-        },
-    }
-    if auth_bridge_mode:
-        spec["authBridgeMode"] = auth_bridge_mode
-    if mtls_mode:
-        spec["mtlsMode"] = mtls_mode
-    # Only set when enabled; unset → operator default "disabled" (also keeps the
-    # CRD field off envoy-sidecar agents so the validating webhook doesn't reject).
-    if tls_bridge_enabled:
-        spec["tlsBridgeMode"] = "enabled"
-    return {
-        "apiVersion": f"{CRD_GROUP}/{CRD_VERSION}",
-        "kind": "AgentRuntime",
-        "metadata": {
-            "name": name,
-            "namespace": namespace,
-            "labels": {
-                ROSSOCTL_TYPE_LABEL: agent_type,
-                APP_KUBERNETES_IO_MANAGED_BY: ROSSOCTL_UI_CREATOR_LABEL,
-            },
-        },
-        "spec": spec,
-    }
-
-
-def _ensure_agentruntime(
-    kube: "KubernetesService",
-    name: str,
-    namespace: str,
-    workload_type: str,
-    agent_type: str = RESOURCE_TYPE_AGENT,
-    auth_bridge_mode: Optional[str] = None,
-    mtls_mode: Optional[str] = None,
-    tls_bridge_enabled: bool = False,
-) -> None:
-    """Create an AgentRuntime CR for the workload. Skip if it already exists."""
-    manifest = _build_agentruntime_manifest(
-        name,
-        namespace,
-        workload_type,
-        agent_type,
-        auth_bridge_mode,
-        mtls_mode,
-        tls_bridge_enabled,
-    )
-    try:
-        kube.create_custom_resource(
-            group=CRD_GROUP,
-            version=CRD_VERSION,
-            namespace=namespace,
-            plural=AGENTRUNTIMES_PLURAL,
-            body=manifest,
-        )
-        logger.info("Created AgentRuntime '%s' in namespace '%s'", name, namespace)
-    except ApiException as e:
-        if e.status == 409:
-            logger.info("AgentRuntime '%s' already exists in namespace '%s'", name, namespace)
-        else:
-            logger.warning("Failed to create AgentRuntime '%s': %s", name, e.reason)
-
-
-def _build_deployment_manifest(
-    request: "CreateAgentRequest",
-    image: str,
-    shipwright_build_name: Optional[str] = None,
-    local_skills: Optional[List[str]] = None,
-    ext_init_containers: Optional[List[Dict[str, Any]]] = None,
-    ext_volumes: Optional[List[Dict[str, Any]]] = None,
-    ext_volume_mounts: Optional[List[Dict[str, Any]]] = None,
-    ext_skill_paths: Optional[List[str]] = None,
-) -> dict:
-    """
-    Build a Kubernetes Deployment manifest for an agent.
-
-    Args:
-        request: The agent creation request.
-        image: The container image URL.
-        shipwright_build_name: Optional name of the Shipwright Build that created
-            this agent (for annotation tracking).
-
-    Returns:
-        Deployment manifest dictionary.
-    """
-    ext_init_containers = ext_init_containers or []
-    ext_volumes = ext_volumes or []
-    ext_volume_mounts = ext_volume_mounts or []
-    ext_skill_paths = ext_skill_paths or []
-    env_vars = _build_env_vars(request, local_skills=local_skills, ext_skill_paths=ext_skill_paths)
-    skill_volumes, skill_volume_mounts, _ = _get_linked_skill_mounts(
-        request, skills_override=local_skills
-    )
-    labels = _build_common_labels(request, WORKLOAD_TYPE_DEPLOYMENT)
-    selector_labels = _build_selector_labels(request)
-
-    # Build annotations
-    annotations: Dict[str, str] = {
-        ROSSOCTL_DESCRIPTION_ANNOTATION: f"Agent '{request.name}' deployed from UI.",
-    }
-    if request.skills:
-        annotations[AGENT_SKILLS_ANNOTATION] = json.dumps(request.skills)
-    if shipwright_build_name:
-        annotations["rossoctl.io/shipwright-build"] = shipwright_build_name
-
-    # Build container ports
-    container_port = DEFAULT_IN_CLUSTER_PORT
-    if request.servicePorts and len(request.servicePorts) > 0:
-        container_port = request.servicePorts[0].targetPort
-
-    manifest = {
-        "apiVersion": "apps/v1",
-        "kind": "Deployment",
-        "metadata": {
-            "name": request.name,
-            "namespace": request.namespace,
-            "labels": labels,
-            "annotations": annotations,
-        },
-        "spec": {
-            "replicas": 1,
-            "selector": {
-                "matchLabels": selector_labels,
-            },
-            "template": {
-                "metadata": {
-                    "labels": {
-                        **labels,
-                    },
-                    "annotations": _build_common_annotations(request),
-                },
-                "spec": {
-                    "serviceAccountName": request.name,
-                    "containers": [
-                        {
-                            "name": "agent",
-                            "image": image,
-                            "imagePullPolicy": DEFAULT_IMAGE_POLICY,
-                            "resources": build_container_resources(
-                                request.k8sResourceLimits, request.k8sResourceRequests
-                            ),
-                            "env": env_vars,
-                            "ports": [
-                                {
-                                    "name": "http",
-                                    "containerPort": container_port,
-                                    "protocol": "TCP",
-                                },
-                            ],
-                            "volumeMounts": [
-                                {"name": "cache", "mountPath": "/app/.cache"},
-                                {"name": "marvin", "mountPath": "/.marvin"},
-                                {"name": "shared-data", "mountPath": "/shared"},
-                                *skill_volume_mounts,
-                                *ext_volume_mounts,
-                            ],
-                        }
-                    ],
-                    "volumes": [
-                        {"name": "cache", "emptyDir": {}},
-                        {"name": "marvin", "emptyDir": {}},
-                        {"name": "shared-data", "emptyDir": {}},
-                        *skill_volumes,
-                        *ext_volumes,
-                    ],
-                },
-            },
-        },
-    }
-
-    # Add init containers for external skills
-    if ext_init_containers:
-        manifest["spec"]["template"]["spec"]["initContainers"] = ext_init_containers
-
-    # Add image pull secrets if specified
-    if request.imagePullSecret:
-        manifest["spec"]["template"]["spec"]["imagePullSecrets"] = [
-            {"name": request.imagePullSecret}
-        ]
-
-    return manifest
-
-
-def _create_or_replace_service(
-    kube: "KubernetesService",
-    namespace: str,
-    name: str,
-    service_manifest: dict,
-    workload_type: str,
-) -> None:
-    """Create the Service for an agent / tool.
-
-    Returns silently for ``WORKLOAD_TYPE_JOB`` since Jobs don't need a Service.
-    Sandbox agents get a backend-managed ClusterIP Service for port translation
-    (8080→8000); the agent-sandbox controller's own Service is suppressed via
-    ``spec.service: false`` on the Sandbox CR (v0.4.6+).
-    """
-    if workload_type == WORKLOAD_TYPE_JOB:
-        return
-    # Strip CR/LF before logging — name and namespace come from the FastAPI
-    # request body. Kubernetes will reject non-DNS-1123 names so this is
-    # belt-and-suspenders, but the explicit sanitization satisfies CodeQL's
-    # py/log-injection taint analysis on the user-input → log-sink flow.
-    safe_name = name.replace("\n", "").replace("\r", "")
-    safe_namespace = namespace.replace("\n", "").replace("\r", "")
-    kube.create_service(namespace=namespace, body=service_manifest)
-    logger.info("Created Service '%s' in namespace '%s'", safe_name, safe_namespace)
-
-
-def _build_service_manifest(request: "CreateAgentRequest") -> dict:
-    """
-    Build a Kubernetes Service manifest for an agent.
-
-    Args:
-        request: The agent creation request.
-
-    Returns:
-        Service manifest dictionary.
-    """
-    labels = _build_common_labels(request, request.workloadType)
-    selector_labels = _build_selector_labels(request)
-
-    # Build service ports
-    if request.servicePorts:
-        service_ports = [
-            {
-                "name": sp.name,
-                "port": sp.port,
-                "targetPort": sp.targetPort,
-                "protocol": sp.protocol,
-            }
-            for sp in request.servicePorts
-        ]
-    else:
-        service_ports = [
-            {
-                "name": "http",
-                "port": DEFAULT_OFF_CLUSTER_PORT,
-                "targetPort": DEFAULT_IN_CLUSTER_PORT,
-                "protocol": "TCP",
-            }
-        ]
-
-    return {
-        "apiVersion": "v1",
-        "kind": "Service",
-        "metadata": {
-            "name": request.name,
-            "namespace": request.namespace,
-            "labels": labels,
-        },
-        "spec": {
-            "type": "ClusterIP",
-            "selector": selector_labels,
-            "ports": service_ports,
-        },
-    }
-
-
-def _build_statefulset_manifest(
-    request: "CreateAgentRequest",
-    image: str,
-    shipwright_build_name: Optional[str] = None,
-    local_skills: Optional[List[str]] = None,
-    ext_init_containers: Optional[List[Dict[str, Any]]] = None,
-    ext_volumes: Optional[List[Dict[str, Any]]] = None,
-    ext_volume_mounts: Optional[List[Dict[str, Any]]] = None,
-    ext_skill_paths: Optional[List[str]] = None,
-) -> dict:
-    """
-    Build a Kubernetes StatefulSet manifest for an agent.
-
-    StatefulSets are useful for agents that require:
-    - Stable, unique network identifiers
-    - Stable, persistent storage
-    - Ordered, graceful deployment and scaling
-    - Ordered, automated rolling updates
-
-    Args:
-        request: The agent creation request.
-        image: The container image URL.
-        shipwright_build_name: Optional name of the Shipwright Build.
-
-    Returns:
-        StatefulSet manifest dictionary.
-    """
-    ext_init_containers = ext_init_containers or []
-    ext_volumes = ext_volumes or []
-    ext_volume_mounts = ext_volume_mounts or []
-    ext_skill_paths = ext_skill_paths or []
-    env_vars = _build_env_vars(request, local_skills=local_skills, ext_skill_paths=ext_skill_paths)
-    skill_volumes, skill_volume_mounts, _ = _get_linked_skill_mounts(
-        request, skills_override=local_skills
-    )
-    labels = _build_common_labels(request, WORKLOAD_TYPE_STATEFULSET)
-    selector_labels = _build_selector_labels(request)
-
-    # Build annotations
-    annotations: Dict[str, str] = {
-        ROSSOCTL_DESCRIPTION_ANNOTATION: f"Agent '{request.name}' deployed as StatefulSet from UI.",
-    }
-    if request.skills:
-        annotations[AGENT_SKILLS_ANNOTATION] = json.dumps(request.skills)
-    if shipwright_build_name:
-        annotations["rossoctl.io/shipwright-build"] = shipwright_build_name
-
-    # Build container ports
-    container_port = DEFAULT_IN_CLUSTER_PORT
-    if request.servicePorts and len(request.servicePorts) > 0:
-        container_port = request.servicePorts[0].targetPort
-
-    manifest = {
-        "apiVersion": "apps/v1",
-        "kind": "StatefulSet",
-        "metadata": {
-            "name": request.name,
-            "namespace": request.namespace,
-            "labels": labels,
-            "annotations": annotations,
-        },
-        "spec": {
-            "serviceName": request.name,  # StatefulSet requires a headless service name
-            "replicas": 1,
-            "selector": {
-                "matchLabels": selector_labels,
-            },
-            "template": {
-                "metadata": {
-                    "labels": {
-                        **labels,
-                    },
-                    "annotations": _build_common_annotations(request),
-                },
-                "spec": {
-                    "serviceAccountName": request.name,
-                    "containers": [
-                        {
-                            "name": "agent",
-                            "image": image,
-                            "imagePullPolicy": DEFAULT_IMAGE_POLICY,
-                            "resources": build_container_resources(
-                                request.k8sResourceLimits, request.k8sResourceRequests
-                            ),
-                            "env": env_vars,
-                            "ports": [
-                                {
-                                    "name": "http",
-                                    "containerPort": container_port,
-                                    "protocol": "TCP",
-                                },
-                            ],
-                            "volumeMounts": [
-                                {"name": "cache", "mountPath": "/app/.cache"},
-                                {"name": "marvin", "mountPath": "/.marvin"},
-                                {"name": "shared-data", "mountPath": "/shared"},
-                                *skill_volume_mounts,
-                                *ext_volume_mounts,
-                            ],
-                        }
-                    ],
-                    "volumes": [
-                        {"name": "cache", "emptyDir": {}},
-                        {"name": "marvin", "emptyDir": {}},
-                        *skill_volumes,
-                        *ext_volumes,
-                    ]
-                    + (
-                        []
-                        if request.persistentStorage and request.persistentStorage.enabled
-                        else [{"name": "shared-data", "emptyDir": {}}]
-                    ),
-                },
-            },
-        },
-    }
-
-    # Add init containers for external skills
-    if ext_init_containers:
-        manifest["spec"]["template"]["spec"]["initContainers"] = ext_init_containers
-
-    # When persistent storage is requested, declare a volumeClaimTemplate so the
-    # StatefulSet provisions a PVC bound to the pod's stable identity; the
-    # shared-data volume above is omitted from `volumes` in that case because
-    # the template name (shared-data) becomes the volume.
-    if request.persistentStorage and request.persistentStorage.enabled:
-        manifest["spec"]["volumeClaimTemplates"] = [
-            {
-                "metadata": {
-                    "name": "shared-data",
-                    "labels": {APP_KUBERNETES_IO_NAME: request.name},
-                },
-                "spec": {
-                    "accessModes": ["ReadWriteOnce"],
-                    "resources": {"requests": {"storage": request.persistentStorage.size}},
-                },
-            }
-        ]
-
-    # Add image pull secrets if specified
-    if request.imagePullSecret:
-        manifest["spec"]["template"]["spec"]["imagePullSecrets"] = [
-            {"name": request.imagePullSecret}
-        ]
-
-    return manifest
-
-
-def _build_job_manifest(
-    request: "CreateAgentRequest",
-    image: str,
-    shipwright_build_name: Optional[str] = None,
-    local_skills: Optional[List[str]] = None,
-    ext_init_containers: Optional[List[Dict[str, Any]]] = None,
-    ext_volumes: Optional[List[Dict[str, Any]]] = None,
-    ext_volume_mounts: Optional[List[Dict[str, Any]]] = None,
-    ext_skill_paths: Optional[List[str]] = None,
-) -> dict:
-    """
-    Build a Kubernetes Job manifest for an agent.
-
-    Jobs are useful for agents that:
-    - Run to completion (batch processing)
-    - Should not be restarted automatically
-    - Perform one-time tasks or scheduled workloads
-
-    Args:
-        request: The agent creation request.
-        image: The container image URL.
-        shipwright_build_name: Optional name of the Shipwright Build.
-
-    Returns:
-        Job manifest dictionary.
-    """
-    ext_init_containers = ext_init_containers or []
-    ext_volumes = ext_volumes or []
-    ext_volume_mounts = ext_volume_mounts or []
-    ext_skill_paths = ext_skill_paths or []
-    env_vars = _build_env_vars(request, local_skills=local_skills, ext_skill_paths=ext_skill_paths)
-    skill_volumes, skill_volume_mounts, _ = _get_linked_skill_mounts(
-        request, skills_override=local_skills
-    )
-    labels = _build_common_labels(request, WORKLOAD_TYPE_JOB)
-
-    # Build annotations
-    annotations: Dict[str, str] = {
-        ROSSOCTL_DESCRIPTION_ANNOTATION: f"Agent '{request.name}' deployed as Job from UI.",
-    }
-    if request.skills:
-        annotations[AGENT_SKILLS_ANNOTATION] = json.dumps(request.skills)
-    if shipwright_build_name:
-        annotations["rossoctl.io/shipwright-build"] = shipwright_build_name
-
-    # Build container ports
-    container_port = DEFAULT_IN_CLUSTER_PORT
-    if request.servicePorts and len(request.servicePorts) > 0:
-        container_port = request.servicePorts[0].targetPort
-
-    manifest = {
-        "apiVersion": "batch/v1",
-        "kind": "Job",
-        "metadata": {
-            "name": request.name,
-            "namespace": request.namespace,
-            "labels": labels,
-            "annotations": annotations,
-        },
-        "spec": {
-            "backoffLimit": 3,  # Number of retries before considering the job failed
-            "template": {
-                "metadata": {
-                    "labels": {
-                        **labels,
-                    },
-                    "annotations": _build_common_annotations(request),
-                },
-                "spec": {
-                    "serviceAccountName": request.name,
-                    "restartPolicy": "OnFailure",
-                    "containers": [
-                        {
-                            "name": "agent",
-                            "image": image,
-                            "imagePullPolicy": DEFAULT_IMAGE_POLICY,
-                            "resources": build_container_resources(
-                                request.k8sResourceLimits, request.k8sResourceRequests
-                            ),
-                            "env": env_vars,
-                            "ports": [
-                                {
-                                    "name": "http",
-                                    "containerPort": container_port,
-                                    "protocol": "TCP",
-                                },
-                            ],
-                            "volumeMounts": [
-                                {"name": "cache", "mountPath": "/app/.cache"},
-                                {"name": "marvin", "mountPath": "/.marvin"},
-                                {"name": "shared-data", "mountPath": "/shared"},
-                                *skill_volume_mounts,
-                                *ext_volume_mounts,
-                            ],
-                        }
-                    ],
-                    "volumes": [
-                        {"name": "cache", "emptyDir": {}},
-                        {"name": "marvin", "emptyDir": {}},
-                        {"name": "shared-data", "emptyDir": {}},
-                        *skill_volumes,
-                        *ext_volumes,
-                    ],
-                },
-            },
-        },
-    }
-
-    # Add init containers for external skills
-    if ext_init_containers:
-        manifest["spec"]["template"]["spec"]["initContainers"] = ext_init_containers
-
-    # Add image pull secrets if specified
-    if request.imagePullSecret:
-        manifest["spec"]["template"]["spec"]["imagePullSecrets"] = [
-            {"name": request.imagePullSecret}
-        ]
-
-    return manifest
-
-
-def _build_sandbox_manifest(
-    request: "CreateAgentRequest",
-    image: str,
-    shipwright_build_name: Optional[str] = None,
-    local_skills: Optional[List[str]] = None,
-    ext_init_containers: Optional[List[Dict[str, Any]]] = None,
-    ext_volumes: Optional[List[Dict[str, Any]]] = None,
-    ext_volume_mounts: Optional[List[Dict[str, Any]]] = None,
-    ext_skill_paths: Optional[List[str]] = None,
-) -> dict:
-    """Build a Sandbox manifest (agents.x-k8s.io/v1alpha1) for direct creation.
-
-    Includes skill volume mounts and persistent storage support.
-    """
-    ext_init_containers = ext_init_containers or []
-    ext_volumes = ext_volumes or []
-    ext_volume_mounts = ext_volume_mounts or []
-    ext_skill_paths = ext_skill_paths or []
-    env_vars = _build_env_vars(request, local_skills=local_skills, ext_skill_paths=ext_skill_paths)
-    skill_volumes, skill_volume_mounts, _ = _get_linked_skill_mounts(
-        request, skills_override=local_skills
-    )
-    labels = _build_common_labels(request, WORKLOAD_TYPE_SANDBOX)
-
-    annotations: Dict[str, str] = {
-        ROSSOCTL_DESCRIPTION_ANNOTATION: f"Agent '{request.name}' deployed from UI.",
-    }
-    if request.skills:
-        annotations[AGENT_SKILLS_ANNOTATION] = json.dumps(request.skills)
-    if shipwright_build_name:
-        annotations["rossoctl.io/shipwright-build"] = shipwright_build_name
-
-    container_port = DEFAULT_IN_CLUSTER_PORT
-    if request.servicePorts and len(request.servicePorts) > 0:
-        container_port = request.servicePorts[0].targetPort
-
-    manifest = {
-        "apiVersion": f"{AGENT_SANDBOX_CRD_GROUP}/{AGENT_SANDBOX_CRD_VERSION}",
-        "kind": "Sandbox",
-        "metadata": {
-            "name": request.name,
-            "namespace": request.namespace,
-            "labels": labels,
-            "annotations": annotations,
-        },
-        "spec": {
-            "replicas": 1,
-            "service": False,
-            "podTemplate": {
-                "metadata": {
-                    "labels": {
-                        **labels,
-                    },
-                    "annotations": _build_common_annotations(request),
-                },
-                "spec": {
-                    "automountServiceAccountToken": False,
-                    "serviceAccountName": request.name,
-                    "containers": [
-                        {
-                            "name": "agent",
-                            "image": image,
-                            "imagePullPolicy": DEFAULT_IMAGE_POLICY,
-                            "resources": build_container_resources(
-                                request.k8sResourceLimits, request.k8sResourceRequests
-                            ),
-                            "env": env_vars,
-                            "ports": [
-                                {
-                                    "name": "http",
-                                    "containerPort": container_port,
-                                    "protocol": "TCP",
-                                },
-                            ],
-                            "volumeMounts": [
-                                {"name": "cache", "mountPath": "/app/.cache"},
-                                {"name": "marvin", "mountPath": "/.marvin"},
-                                {"name": "shared-data", "mountPath": "/shared"},
-                                *skill_volume_mounts,
-                                *ext_volume_mounts,
-                            ],
-                        }
-                    ],
-                    "volumes": [
-                        {"name": "cache", "emptyDir": {}},
-                        {"name": "marvin", "emptyDir": {}},
-                        *skill_volumes,
-                        *ext_volumes,
-                    ]
-                    + (
-                        []
-                        if request.persistentStorage and request.persistentStorage.enabled
-                        else [{"name": "shared-data", "emptyDir": {}}]
-                    ),
-                },
-            },
-        },
-    }
-
-    # Add init containers for external skills
-    if ext_init_containers:
-        manifest["spec"]["podTemplate"]["spec"]["initContainers"] = ext_init_containers
-
-    if request.persistentStorage and request.persistentStorage.enabled:
-        manifest["spec"]["volumeClaimTemplates"] = [
-            {
-                "metadata": {
-                    "name": "shared-data",
-                    "labels": {APP_KUBERNETES_IO_NAME: request.name},
-                },
-                "spec": {
-                    "accessModes": ["ReadWriteOnce"],
-                    "resources": {"requests": {"storage": request.persistentStorage.size}},
-                },
-            }
-        ]
-
-    if request.imagePullSecret:
-        manifest["spec"]["podTemplate"]["spec"]["imagePullSecrets"] = [
-            {"name": request.imagePullSecret}
-        ]
-
-    return manifest
-
-
-async def _resolve_context_mounts(
-    namespace: str,
-    attachments: Optional[List[ContextAttachment]],
-    workload_type: str,
-) -> Tuple[List[Dict[str, Any]], List[Dict[str, Any]]]:
-    """Resolve named contexts into Kubernetes PVC volumes and mounts."""
-    if not attachments:
-        return [], []
-    if workload_type not in {WORKLOAD_TYPE_STATEFULSET, WORKLOAD_TYPE_SANDBOX}:
-        raise HTTPException(
-            status_code=400,
-            detail="context attachments require a statefulset or sandbox workload",
-        )
-    if not settings.context_service_url.strip():
-        raise HTTPException(status_code=400, detail="Context Service integration is disabled")
-
-    from app.routers.contexts import resolve_context  # pylint: disable=import-outside-toplevel
-
-    volumes: List[Dict[str, Any]] = []
-    mounts: List[Dict[str, Any]] = []
-    seen_paths: set[str] = set()
-    for index, attachment in enumerate(attachments):
-        if not attachment.mountPath.startswith("/"):
-            raise HTTPException(status_code=400, detail="context mountPath must be absolute")
-        if attachment.mountPath in seen_paths:
-            raise HTTPException(status_code=400, detail="context mountPath values must be unique")
-        seen_paths.add(attachment.mountPath)
-        resource = await resolve_context(namespace, attachment.name)
-        resource_attachment = resource.get("attachment", {})
-        if resource_attachment.get("kind") != "pvc":
-            raise HTTPException(
-                status_code=502, detail="Context Service returned a non-PVC attachment"
-            )
-        claim_name = resource_attachment.get("claimName")
-        if not claim_name:
-            raise HTTPException(status_code=502, detail="Context Service returned no PVC claim")
-        volume_name = f"context-{index}"
-        volumes.append(
-            {
-                "name": volume_name,
-                "persistentVolumeClaim": {
-                    "claimName": claim_name,
-                    "readOnly": attachment.readOnly,
-                },
-            }
-        )
-        mounts.append(
-            {
-                "name": volume_name,
-                "mountPath": attachment.mountPath,
-                "readOnly": attachment.readOnly,
-            }
-        )
-    return volumes, mounts
-
-
-@router.post(
+# Sub-routers are composed here, AFTER the four endpoints defined above, so that
+# registration order matches the original single-file layout. Order is
+# significant: FastAPI resolves routes in registration order, and these are the
+# positions the endpoints occupied before the split.
+#
+# `list_agents` deliberately stays in this module: tests patch
+# "app.routers.agents.settings", which only affects code whose module globals
+# live here. Moving it would silently disable that patch.
+router.include_router(migration_router)
+router.include_router(shipwright_router)
+# `create_agent` is registered directly rather than via a sub-router: its path is
+# "" (exactly /agents), which FastAPI rejects on a prefix-less sub-router.
+router.post(
     "", response_model=CreateAgentResponse, dependencies=[Depends(require_roles(ROLE_OPERATOR))]
-)
-async def create_agent(
-    request: CreateAgentRequest,
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> CreateAgentResponse:
-    """
-    Create a new agent.
-
-    Supports two deployment methods:
-    - 'source': Build from git repository using Shipwright Build + BuildRun
-    - 'image': Deploy from existing container image as workload + Service
-
-    Supports four workload types:
-    - 'deployment': Standard Kubernetes Deployment (default)
-    - 'statefulset': StatefulSet for stateful agents
-    - 'job': Job for batch/one-time agents
-    - 'sandbox': Sandbox CR for isolated agents (requires feature flag)
-    """
-    logger.info(
-        f"Creating agent '{request.name}' in namespace '{request.namespace}', "
-        f"workloadType={request.workloadType}, "
-        f"createHttpRoute={request.createHttpRoute}"
-    )
-
-    # Feature flag: reject skill linking if feature is disabled
-    if request.skills and not settings.rossoctl_feature_flag_skills:
-        raise HTTPException(
-            status_code=400,
-            detail="Skill linking is disabled. Enable ROSSOCTL_FEATURE_FLAG_SKILLS to use this feature.",
-        )
-
-    # Compute external skill data when feature is enabled
-    local_skills: Optional[List[str]] = None
-    ext_init_containers: List[Dict[str, Any]] = []
-    ext_volumes: List[Dict[str, Any]] = []
-    ext_volume_mounts: List[Dict[str, Any]] = []
-    ext_skill_paths: List[str] = []
-
-    if request.skills and settings.rossoctl_feature_flag_external_skills:
-        _ensure_fetcher_scripts_cm(kube, request.namespace)
-        ext_init_containers, ext_volumes, ext_volume_mounts, ext_skill_paths = (
-            _get_external_skill_data(kube, request.namespace, request.skills)
-        )
-        local_skills = [
-            s for s in request.skills if s and not _is_skill_external(kube, request.namespace, s)
-        ]
-
-    context_volumes, context_mounts = await _resolve_context_mounts(
-        request.namespace, request.contexts, request.workloadType
-    )
-    ext_volumes.extend(context_volumes)
-    ext_volume_mounts.extend(context_mounts)
-
-    request = apply_agent_import_defaults(request, kube)
-
-    # Persistent resources created during this call, tracked so we can roll them
-    # back if a later creation step fails (avoids leaking a workload, Service,
-    # AgentRuntime, or route). Only used by the image-deployment path below.
-    created: List[Tuple[str, str]] = []
-    try:
-        if request.deploymentMethod == "image":
-            # Deploy from existing container image
-            if not request.containerImage:
-                raise HTTPException(
-                    status_code=400,
-                    detail="containerImage is required for image deployment",
-                )
-
-            # Ensure a dedicated ServiceAccount exists so the webhook's
-            # SPIFFE identity uses the workload name, not the ReplicaSet hash.
-            kube.ensure_service_account(namespace=request.namespace, name=request.name)
-
-            # Ensure AuthBridge ConfigMaps exist in the target namespace
-            if request.authBridgeEnabled:
-                _ensure_authbridge_configmaps(
-                    kube=kube,
-                    namespace=request.namespace,
-                    spire_enabled=request.spireEnabled,
-                )
-                if request.outboundRoutes:
-                    _ensure_authproxy_routes(
-                        kube=kube,
-                        namespace=request.namespace,
-                        routes=request.outboundRoutes,
-                    )
-                if request.defaultOutboundPolicy:
-                    extra_config = {
-                        "DEFAULT_OUTBOUND_POLICY": request.defaultOutboundPolicy,
-                    }
-                    kube.upsert_configmap(
-                        namespace=request.namespace,
-                        name="authbridge-config",
-                        data=extra_config,
-                    )
-
-            # On OpenShift, ensure the AuthBridge SCC RoleBinding exists
-            if request.authBridgeEnabled:
-                _ensure_authbridge_scc_rolebinding(kube=kube, namespace=request.namespace)
-
-            # Create card-unsigned ConfigMap so the webhook injects
-            # the sign-agentcard init container at Deployment admission.
-            if request.spireEnabled:
-                service_port = (
-                    request.servicePorts[0].port
-                    if request.servicePorts
-                    else DEFAULT_IN_CLUSTER_PORT
-                )
-                _ensure_card_unsigned_configmap(
-                    kube=kube,
-                    name=request.name,
-                    namespace=request.namespace,
-                    service_port=service_port,
-                    description=f"Agent '{request.name}' deployed from UI.",
-                    skill_names=request.skills,
-                )
-
-            # Create workload based on workloadType
-            if request.workloadType == WORKLOAD_TYPE_DEPLOYMENT:
-                workload_manifest = _build_deployment_manifest(
-                    request=request,
-                    image=request.containerImage,
-                    local_skills=local_skills,
-                    ext_init_containers=ext_init_containers,
-                    ext_volumes=ext_volumes,
-                    ext_volume_mounts=ext_volume_mounts,
-                    ext_skill_paths=ext_skill_paths,
-                )
-                kube.create_deployment(
-                    namespace=request.namespace,
-                    body=workload_manifest,
-                )
-                created.append(("Deployment", request.name))
-                logger.info(
-                    f"Created Deployment '{request.name}' in namespace '{request.namespace}'"
-                )
-            elif request.workloadType == WORKLOAD_TYPE_STATEFULSET:
-                workload_manifest = _build_statefulset_manifest(
-                    request=request,
-                    image=request.containerImage,
-                    local_skills=local_skills,
-                    ext_init_containers=ext_init_containers,
-                    ext_volumes=ext_volumes,
-                    ext_volume_mounts=ext_volume_mounts,
-                    ext_skill_paths=ext_skill_paths,
-                )
-                kube.create_statefulset(
-                    namespace=request.namespace,
-                    body=workload_manifest,
-                )
-                created.append(("StatefulSet", request.name))
-                logger.info(
-                    f"Created StatefulSet '{request.name}' in namespace '{request.namespace}'"
-                )
-            elif request.workloadType == WORKLOAD_TYPE_JOB:
-                workload_manifest = _build_job_manifest(
-                    request=request,
-                    image=request.containerImage,
-                    local_skills=local_skills,
-                    ext_init_containers=ext_init_containers,
-                    ext_volumes=ext_volumes,
-                    ext_volume_mounts=ext_volume_mounts,
-                    ext_skill_paths=ext_skill_paths,
-                )
-                kube.create_job(
-                    namespace=request.namespace,
-                    body=workload_manifest,
-                )
-                created.append(("Job", request.name))
-                logger.info(f"Created Job '{request.name}' in namespace '{request.namespace}'")
-            elif request.workloadType == WORKLOAD_TYPE_SANDBOX:
-                sandbox_manifest = _build_sandbox_manifest(
-                    request=request,
-                    image=request.containerImage,
-                    local_skills=local_skills,
-                    ext_init_containers=ext_init_containers,
-                    ext_volumes=ext_volumes,
-                    ext_volume_mounts=ext_volume_mounts,
-                    ext_skill_paths=ext_skill_paths,
-                )
-                kube.create_sandbox(
-                    namespace=request.namespace,
-                    body=sandbox_manifest,
-                )
-                created.append(("Sandbox", request.name))
-                logger.info(f"Created Sandbox '{request.name}' in namespace '{request.namespace}'")
-
-            # Create Service (not needed for Jobs).
-            if request.workloadType != WORKLOAD_TYPE_JOB:
-                service_manifest = _build_service_manifest(request)
-                _create_or_replace_service(
-                    kube,
-                    request.namespace,
-                    request.name,
-                    service_manifest,
-                    request.workloadType,
-                )
-                created.append(("Service", request.name))
-
-            # Create AgentRuntime CR so the per-agent AuthBridge config (mtls /
-            # authBridgeMode / tlsBridgeMode) is applied. Sandbox is included
-            # (targetRef -> agents.x-k8s.io Sandbox); only Job is excluded —
-            # a run-to-completion Job doesn't fit the attach/restart model.
-            if _agentruntime_supported_workload(request.workloadType):
-                _ensure_agentruntime(
-                    kube=kube,
-                    name=request.name,
-                    namespace=request.namespace,
-                    workload_type=request.workloadType,
-                    auth_bridge_mode=request.authBridgeMode,
-                    mtls_mode=request.mtlsMode,
-                    tls_bridge_enabled=request.tlsBridgeEnabled,
-                )
-                created.append(("AgentRuntime", request.name))
-
-            message = f"Agent '{request.name}' deployed as {request.workloadType} successfully."
-
-            # Create HTTPRoute/Route if requested (not applicable for Jobs or Sandboxes)
-            if request.createHttpRoute and request.workloadType not in (
-                WORKLOAD_TYPE_JOB,
-                WORKLOAD_TYPE_SANDBOX,
-            ):
-                service_port = select_route_port(
-                    request.servicePorts,
-                    default_port=DEFAULT_OFF_CLUSTER_PORT,
-                )
-                create_route_for_agent_or_tool(
-                    kube=kube,
-                    name=request.name,
-                    namespace=request.namespace,
-                    service_name=request.name,
-                    service_port=service_port,
-                )
-                # create_route_for_agent_or_tool makes an HTTPRoute or an OpenShift
-                # Route depending on platform; track both so rollback deletes the
-                # right one (the other 404s and is swallowed).
-                created.append(("HTTPRoute", request.name))
-                created.append(("Route", request.name))
-                message += " HTTPRoute/Route created for external access."
-
-        else:
-            # Build from source using Shipwright Build + BuildRun
-            if not request.gitUrl:
-                raise HTTPException(
-                    status_code=400,
-                    detail="gitUrl is required for source deployment",
-                )
-
-            # Clean up any existing Build/BuildRuns to prevent 409 on re-import
-            cleanup_existing_build(kube, namespace=request.namespace, build_name=request.name)
-
-            # Step 1: Create Shipwright Build CR
-            clone_secret = resolve_clone_secret(kube.core_api, request.namespace)
-            build_manifest = _build_agent_shipwright_build_manifest(
-                request, clone_secret_name=clone_secret
-            )
-            kube.create_custom_resource(
-                group=SHIPWRIGHT_CRD_GROUP,
-                version=SHIPWRIGHT_CRD_VERSION,
-                namespace=request.namespace,
-                plural=SHIPWRIGHT_BUILDS_PLURAL,
-                body=build_manifest,
-            )
-            logger.info(
-                f"Created Shipwright Build '{request.name}' in namespace '{request.namespace}'"
-            )
-
-            # Step 2: Create BuildRun CR to trigger the build
-            # Get labels from the Build manifest to propagate to BuildRun
-            build_labels = build_manifest.get("metadata", {}).get("labels", {})
-            buildrun_manifest = _build_agent_shipwright_buildrun_manifest(
-                build_name=request.name,
-                namespace=request.namespace,
-                labels=build_labels,
-            )
-            created_buildrun = kube.create_custom_resource(
-                group=SHIPWRIGHT_CRD_GROUP,
-                version=SHIPWRIGHT_CRD_VERSION,
-                namespace=request.namespace,
-                plural=SHIPWRIGHT_BUILDRUNS_PLURAL,
-                body=buildrun_manifest,
-            )
-            buildrun_name = created_buildrun.get("metadata", {}).get("name", "")
-            logger.info(
-                f"Created Shipwright BuildRun '{buildrun_name}' in namespace '{request.namespace}'"
-            )
-
-            message = (
-                f"Shipwright build started for agent '{request.name}'. "
-                f"BuildRun: '{buildrun_name}'. "
-                f"Poll the build status and create the Agent after the build completes."
-            )
-
-            # Note: For Shipwright builds, HTTPRoute is NOT created here.
-            # It will be created when the Agent is finalized after build completion.
-            if request.createHttpRoute:
-                message += " HTTPRoute will be created after the build completes."
-
-        return CreateAgentResponse(
-            success=True,
-            name=request.name,
-            namespace=request.namespace,
-            message=message,
-        )
-
-    except ApiException as e:
-        # Roll back only what THIS call created (tracked in `created`); if the very
-        # first create 409'd, `created` is empty and rollback is a no-op, so a
-        # pre-existing agent is never deleted.
-        rollback_workload_resources(kube, request.namespace, created)
-        if e.status == 409:
-            raise HTTPException(
-                status_code=409,
-                detail=f"Agent '{request.name}' already exists in namespace '{request.namespace}'",
-            )
-        if e.status == 404:
-            raise HTTPException(
-                status_code=404,
-                detail=(
-                    f"Required CRD or resource not found for workload type "
-                    f"'{request.workloadType}'. Ensure the necessary controllers "
-                    f"are installed (e.g. Shipwright for source builds, "
-                    f"agent-sandbox controller for sandbox workloads)."
-                ),
-            )
-        logger.error(f"Failed to create agent: {e}")
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-    except HTTPException:
-        # Validation errors (400) raised above — nothing created yet, re-raise as-is.
-        raise
-    except Exception as e:
-        # Non-API failure (e.g. platform detection in route creation) after some
-        # resources were already created — roll back before surfacing a 500.
-        rollback_workload_resources(kube, request.namespace, created)
-        logger.error(
-            "Unexpected error creating agent '%s': %s",
-            sanitize_log(request.name),
-            sanitize_log(str(e)),
-        )
-        raise HTTPException(status_code=500, detail=f"Failed to create agent: {e}")
-
-
-class FinalizeShipwrightBuildRequest(BaseModel):
-    """Request to finalize a Shipwright build and create the Agent.
-
-    All fields are optional. If not provided, the values stored in the Build's
-    rossoctl.io/agent-config annotation will be used.
-    """
-
-    # These fields mirror CreateAgentRequest for Agent creation
-    # All optional - will use values from Build annotation if not provided
-    protocol: Optional[str] = None
-    framework: Optional[str] = None
-    envVars: Optional[List[EnvVar]] = None
-    skills: Optional[List[str]] = None
-    servicePorts: Optional[List[ServicePort]] = None
-    createHttpRoute: Optional[bool] = None
-    authBridgeEnabled: Optional[bool] = None
-    imagePullSecret: Optional[str] = None
-    authBridgeMode: Optional[Literal["proxy-sidecar", "envoy-sidecar", "lite", "waypoint"]] = None
-    # Mirrors CreateAgentRequest.mtlsMode. Threaded through the
-    # finalize flow so a build-from-source agent inherits the mtlsMode
-    # the user picked at form-submit time (stashed on the BuildRun via
-    # rossoctl.io/agent-config annotation).
-    mtlsMode: Optional[Literal["disabled", "permissive", "strict"]] = None
-    # Mirrors CreateAgentRequest.tlsBridgeEnabled. None → inherit the value
-    # stashed on the BuildRun annotation at form-submit time.
-    tlsBridgeEnabled: Optional[bool] = None
-    outboundRoutes: Optional[List[OutboundRoute]] = None
-    outboundPortsExclude: Optional[str] = None
-    inboundPortsExclude: Optional[str] = None
-    defaultOutboundPolicy: Optional[Literal["passthrough", "exchange"]] = None
-    persistentStorage: Optional[PersistentStorageConfig] = None
-    mcpToolName: Optional[str] = None
-    llmPreset: Optional[str] = None
-    llmModel: Optional[str] = None
-    # Mirror CreateAgentRequest.k8sResourceLimits/k8sResourceRequests. None →
-    # inherit the value stashed on the BuildRun annotation at form-submit time,
-    # so a build-from-source agent gets the same resources as a direct-image one.
-    k8sResourceLimits: Optional[Dict[str, str]] = None
-    k8sResourceRequests: Optional[Dict[str, str]] = None
-
-    @model_validator(mode="after")
-    def _check_mtls_compatible_with_mode(self) -> "FinalizeShipwrightBuildRequest":
-        """Mirror of CreateAgentRequest._check_mtls_compatible_with_mode
-        at the Shipwright finalize boundary. Today there are no
-        rejected combinations.
-
-        TODO(future-incompatibility): re-enable cross-field rejections
-        here when a new authBridgeMode lands that needs different mTLS
-        semantics. See CreateAgentRequest._check_mtls_compatible_with_mode
-        for the full rationale (including why SPIRE-vs-mTLS coupling
-        is handled at the operator data-plane layer rather than here).
-        """
-        return self
-
-    @model_validator(mode="after")
-    def _check_tlsbridge_compatible_with_mode(self) -> "FinalizeShipwrightBuildRequest":
-        """Mirror of CreateAgentRequest._check_tlsbridge_compatible_with_mode at
-        the Shipwright finalize boundary, so a direct finalize caller (or a combo
-        inherited from the BuildRun's stored config) with tlsBridgeEnabled +
-        envoy-sidecar/waypoint gets the same fast 422 instead of a later webhook
-        denial. Same allowlist as the operator (empty → defaults to proxy-sidecar).
-        """
-        allowed = (None, "", "proxy-sidecar", "lite")
-        if self.tlsBridgeEnabled and self.authBridgeMode not in allowed:
-            raise ValueError(
-                "tlsBridgeEnabled requires authBridgeMode proxy-sidecar or lite "
-                f"(the TLS bridge lives in the Go forward proxy); got {self.authBridgeMode!r}"
-            )
-        return self
-
-
-@router.post(
-    "/{namespace}/{name}/finalize-shipwright-build",
-    response_model=CreateAgentResponse,
-    dependencies=[Depends(require_roles(ROLE_OPERATOR))],
-)
-async def finalize_shipwright_build(
-    namespace: str,
-    name: str,
-    request: FinalizeShipwrightBuildRequest,
-    kube: KubernetesService = Depends(get_kubernetes_service),
-) -> CreateAgentResponse:
-    """
-    Finalize a Shipwright build by creating the Deployment and Service.
-
-    This endpoint should be called after the Shipwright BuildRun completes successfully.
-    It retrieves the output image from the BuildRun status and creates the Deployment
-    and Service for the agent.
-
-    Agent configuration can be provided in the request body, or it will be read from
-    the Build's rossoctl.io/agent-config annotation (stored during build creation).
-    """
-    logger.info(f"Finalizing Shipwright build '{name}' in namespace '{namespace}'")
-
-    try:
-        # Step 1: Get the latest BuildRun status to get the output image
-        items = kube.list_custom_resources(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            namespace=namespace,
-            plural=SHIPWRIGHT_BUILDRUNS_PLURAL,
-            label_selector=f"rossoctl.io/build-name={name}",
-        )
-
-        if not items:
-            raise HTTPException(
-                status_code=404,
-                detail=f"No BuildRuns found for build '{name}' in namespace '{namespace}'",
-            )
-
-        # Sort by creation timestamp and get the most recent
-        items.sort(
-            key=lambda x: x.get("metadata", {}).get("creationTimestamp", ""),
-            reverse=True,
-        )
-        latest_buildrun = items[0]
-        buildrun_status = latest_buildrun.get("status", {})
-
-        # Check if build succeeded
-        conditions = buildrun_status.get("conditions") or []
-        build_succeeded = False
-        failure_message = None
-        for cond in conditions:
-            if cond.get("type") == "Succeeded":
-                if cond.get("status") == "True":
-                    build_succeeded = True
-                else:
-                    failure_message = cond.get("message", "Build failed")
-                break
-
-        if not build_succeeded:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Build has not succeeded yet. Status: {failure_message or 'In progress'}",
-            )
-
-        # Get Build resource for labels and stored agent config (needed for workload type check)
-        build = kube.get_custom_resource(
-            group=SHIPWRIGHT_CRD_GROUP,
-            version=SHIPWRIGHT_CRD_VERSION,
-            namespace=namespace,
-            plural=SHIPWRIGHT_BUILDS_PLURAL,
-            name=name,
-        )
-        build_metadata = build.get("metadata", {})
-        build_labels = build_metadata.get("labels", {})
-        build_annotations = build_metadata.get("annotations", {})
-
-        # Parse stored agent config from Build annotations
-        stored_config: Dict[str, Any] = {}
-        agent_config_json = build_annotations.get("rossoctl.io/agent-config")
-        if agent_config_json:
-            try:
-                stored_config = json.loads(agent_config_json)
-            except json.JSONDecodeError as e:
-                logger.warning(f"Failed to parse agent config from Build annotation: {e}")
-
-        # Determine expected workload type from stored config
-        expected_workload_type = stored_config.get("workloadType", WORKLOAD_TYPE_DEPLOYMENT)
-
-        # Check if workload already exists (idempotency check)
-        # This handles the case where finalize is called multiple times
-        workload_exists = False
-        existing_workload_type = None
-        try:
-            kube.get_deployment(namespace=namespace, name=name)
-            workload_exists = True
-            existing_workload_type = WORKLOAD_TYPE_DEPLOYMENT
-        except ApiException as e:
-            if e.status != 404:
-                raise
-        if not workload_exists:
-            try:
-                kube.get_statefulset(namespace=namespace, name=name)
-                workload_exists = True
-                existing_workload_type = WORKLOAD_TYPE_STATEFULSET
-            except ApiException as e:
-                if e.status != 404:
-                    raise
-        if not workload_exists:
-            try:
-                kube.get_job(namespace=namespace, name=name)
-                workload_exists = True
-                existing_workload_type = WORKLOAD_TYPE_JOB
-            except ApiException as e:
-                if e.status != 404:
-                    raise
-        if not workload_exists and settings.rossoctl_feature_flag_agent_sandbox:
-            try:
-                kube.get_sandbox(namespace=namespace, name=name)
-                workload_exists = True
-                existing_workload_type = WORKLOAD_TYPE_SANDBOX
-            except ApiException as e:
-                if e.status != 404:
-                    raise
-
-        if workload_exists:
-            # Check if existing workload type matches expected type from config
-            if existing_workload_type != expected_workload_type:
-                logger.warning(
-                    f"Workload type mismatch for '{name}' in namespace '{namespace}': "
-                    f"existing workload is {existing_workload_type}, but stored config "
-                    f"specifies {expected_workload_type}. This may indicate a configuration issue."
-                )
-                return CreateAgentResponse(
-                    success=True,
-                    name=name,
-                    namespace=namespace,
-                    message=(
-                        f"Agent '{name}' already deployed as {existing_workload_type}, "
-                        f"but stored config specifies {expected_workload_type}. "
-                        "The existing workload was preserved."
-                    ),
-                )
-            logger.info(
-                f"Workload '{name}' already exists as {existing_workload_type} in namespace '{namespace}'. "
-                "Skipping creation (finalize already completed)."
-            )
-            return CreateAgentResponse(
-                success=True,
-                name=name,
-                namespace=namespace,
-                message=f"Agent '{name}' already deployed as {existing_workload_type}.",
-            )
-
-        # Get the output image from BuildRun status
-        output = buildrun_status.get("output", {})
-        output_image = output.get("image")
-        output_digest = output.get("digest")
-
-        if not output_image:
-            # Fallback: try to get image from Build spec (build already fetched earlier)
-            output_image = build.get("spec", {}).get("output", {}).get("image")
-
-        if not output_image:
-            raise HTTPException(
-                status_code=500,
-                detail="Could not determine output image from build",
-            )
-
-        # If we have a digest, use it for immutable image reference
-        container_image = f"{output_image}@{output_digest}" if output_digest else output_image
-
-        # Merge request with stored config (request values take precedence)
-        # Note: build, build_labels, build_annotations, and stored_config were fetched earlier
-        final_protocol = (
-            request.protocol
-            if request.protocol is not None
-            else stored_config.get("protocol", "a2a")
-        )
-        final_framework = (
-            request.framework
-            if request.framework is not None
-            else stored_config.get("framework", "LangGraph")
-        )
-        final_create_route = (
-            request.createHttpRoute
-            if request.createHttpRoute is not None
-            else stored_config.get("createHttpRoute", False)
-        )
-        final_registry_secret = (
-            request.imagePullSecret
-            if request.imagePullSecret is not None
-            else stored_config.get("registrySecret")
-        )
-        final_auth_bridge = (
-            request.authBridgeEnabled
-            if request.authBridgeEnabled is not None
-            else stored_config.get("authBridgeEnabled", True)
-        )
-        # Use expected_workload_type computed earlier (from stored config)
-        final_workload_type = expected_workload_type
-
-        # For envVars and servicePorts, use request if provided, otherwise use stored config
-        final_env_vars = request.envVars
-        if final_env_vars is None and "envVars" in stored_config:
-            # Convert stored dict format back to EnvVar objects
-            final_env_vars = [EnvVar(**ev) for ev in stored_config["envVars"]]
-
-        final_skills = request.skills
-        if final_skills is None:
-            final_skills = stored_config.get("skills")
-
-        # Feature flag: reject skill linking if feature is disabled
-        if final_skills and not settings.rossoctl_feature_flag_skills:
-            raise HTTPException(
-                status_code=400,
-                detail="Skill linking is disabled. Enable ROSSOCTL_FEATURE_FLAG_SKILLS to use this feature.",
-            )
-
-        # Compute external skill data when feature is enabled (build path)
-        build_local_skills: Optional[List[str]] = None
-        build_ext_init_containers: List[Dict[str, Any]] = []
-        build_ext_volumes: List[Dict[str, Any]] = []
-        build_ext_volume_mounts: List[Dict[str, Any]] = []
-        build_ext_skill_paths: List[str] = []
-
-        if final_skills and settings.rossoctl_feature_flag_external_skills:
-            _ensure_fetcher_scripts_cm(kube, namespace)
-            (
-                build_ext_init_containers,
-                build_ext_volumes,
-                build_ext_volume_mounts,
-                build_ext_skill_paths,
-            ) = _get_external_skill_data(kube, namespace, final_skills)
-            build_local_skills = [
-                s for s in final_skills if s and not _is_skill_external(kube, namespace, s)
-            ]
-
-        final_service_ports = request.servicePorts
-        if final_service_ports is None and "servicePorts" in stored_config:
-            # Convert stored dict format back to ServicePort objects
-            final_service_ports = [ServicePort(**sp) for sp in stored_config["servicePorts"]]
-
-        # Propagate SPIRE identity setting from stored config
-        final_spire_enabled = stored_config.get("spireEnabled", False)
-
-        # Port exclusion and advanced config
-        final_outbound_ports_exclude = (
-            request.outboundPortsExclude
-            if request.outboundPortsExclude is not None
-            else stored_config.get("outboundPortsExclude")
-        )
-        final_inbound_ports_exclude = (
-            request.inboundPortsExclude
-            if request.inboundPortsExclude is not None
-            else stored_config.get("inboundPortsExclude")
-        )
-        final_default_outbound_policy = (
-            request.defaultOutboundPolicy
-            if request.defaultOutboundPolicy is not None
-            else stored_config.get("defaultOutboundPolicy")
-        )
-        # Outbound routing rules
-        final_outbound_routes = None
-        stored_routes = stored_config.get("outboundRoutes")
-        if request.outboundRoutes is not None:
-            final_outbound_routes = request.outboundRoutes
-        elif stored_routes:
-            final_outbound_routes = [OutboundRoute(**r) for r in stored_routes]
-
-        # Per-workload AuthBridge mode override
-        final_auth_bridge_mode = (
-            request.authBridgeMode
-            if request.authBridgeMode is not None
-            else stored_config.get("authBridgeMode")
-        )
-
-        # Per-workload mTLS mode (applies to AgentRuntime spec only;
-        # the form stores it on the BuildRun annotation at submit time
-        # and we read it back here so build-from-source agents inherit
-        # the same setting as direct-image agents).
-        final_mtls_mode = (
-            request.mtlsMode if request.mtlsMode is not None else stored_config.get("mtlsMode")
-        )
-
-        # Per-workload TLS bridge (bool; None on the finalize request → inherit
-        # the stored value). Same store-then-read-back flow as mtlsMode.
-        final_tls_bridge_enabled = (
-            request.tlsBridgeEnabled
-            if request.tlsBridgeEnabled is not None
-            else bool(stored_config.get("tlsBridgeEnabled"))
-        )
-
-        # Persistent storage
-        final_persistent_storage = request.persistentStorage
-        if final_persistent_storage is None and stored_config.get("persistentStorage"):
-            final_persistent_storage = PersistentStorageConfig(**stored_config["persistentStorage"])
-
-        # Per-workload resource overrides. Same store-then-read-back flow as
-        # mtlsMode: None on the finalize request → inherit whatever the form
-        # stashed on the BuildRun annotation, else fall back to the platform
-        # defaults inside the manifest builders.
-        final_k8s_resource_limits = (
-            request.k8sResourceLimits
-            if request.k8sResourceLimits is not None
-            else stored_config.get("k8sResourceLimits")
-        )
-        final_k8s_resource_requests = (
-            request.k8sResourceRequests
-            if request.k8sResourceRequests is not None
-            else stored_config.get("k8sResourceRequests")
-        )
-
-        final_contexts = request.contexts
-        if final_contexts is None and stored_config.get("contexts"):
-            final_contexts = [ContextAttachment(**item) for item in stored_config["contexts"]]
-
-        final_mcp_tool_name = (
-            request.mcpToolName
-            if request.mcpToolName is not None
-            else stored_config.get("mcpToolName")
-        )
-        final_llm_preset = (
-            request.llmPreset if request.llmPreset is not None else stored_config.get("llmPreset")
-        )
-        final_llm_model = (
-            request.llmModel if request.llmModel is not None else stored_config.get("llmModel")
-        )
-
-        # Step 3: Create workload + Service with the built image
-        # Build a CreateAgentRequest-like object for manifest builders
-        agent_request = CreateAgentRequest(
-            name=name,
-            namespace=namespace,
-            protocol=final_protocol,
-            framework=final_framework,
-            deploymentMethod="image",
-            workloadType=final_workload_type,
-            containerImage=container_image,
-            imagePullSecret=final_registry_secret,
-            envVars=final_env_vars,
-            skills=final_skills,
-            servicePorts=final_service_ports,
-            createHttpRoute=final_create_route,
-            authBridgeEnabled=final_auth_bridge,
-            spireEnabled=final_spire_enabled,
-            authBridgeMode=final_auth_bridge_mode,
-            mtlsMode=final_mtls_mode,
-            tlsBridgeEnabled=final_tls_bridge_enabled,
-            outboundRoutes=final_outbound_routes,
-            outboundPortsExclude=final_outbound_ports_exclude,
-            inboundPortsExclude=final_inbound_ports_exclude,
-            defaultOutboundPolicy=final_default_outbound_policy,
-            persistentStorage=final_persistent_storage,
-            contexts=final_contexts,
-            gitPath=stored_config.get("gitPath")
-            or build.get("spec", {}).get("source", {}).get("contextDir", ""),
-            mcpToolName=final_mcp_tool_name,
-            llmPreset=final_llm_preset,
-            llmModel=final_llm_model,
-            k8sResourceLimits=final_k8s_resource_limits,
-            k8sResourceRequests=final_k8s_resource_requests,
-        )
-        agent_request = apply_agent_import_defaults(agent_request, kube)
-        context_volumes, context_mounts = await _resolve_context_mounts(
-            namespace, final_contexts, final_workload_type
-        )
-        build_ext_volumes.extend(context_volumes)
-        build_ext_volume_mounts.extend(context_mounts)
-
-        # Ensure a dedicated ServiceAccount exists so the webhook's
-        # SPIFFE identity uses the workload name, not the ReplicaSet hash.
-        kube.ensure_service_account(namespace=namespace, name=name)
-
-        # Ensure AuthBridge ConfigMaps exist in the target namespace
-        if final_auth_bridge:
-            _ensure_authbridge_configmaps(
-                kube=kube,
-                namespace=namespace,
-                spire_enabled=final_spire_enabled,
-            )
-            if final_outbound_routes:
-                _ensure_authproxy_routes(
-                    kube=kube,
-                    namespace=namespace,
-                    routes=final_outbound_routes,
-                )
-
-        # On OpenShift, ensure the AuthBridge SCC RoleBinding exists
-        if final_auth_bridge:
-            _ensure_authbridge_scc_rolebinding(kube=kube, namespace=namespace)
-
-        # Create card-unsigned ConfigMap so the webhook injects
-        # the sign-agentcard init container at Deployment admission.
-        if final_spire_enabled:
-            service_port = (
-                final_service_ports[0].port if final_service_ports else DEFAULT_IN_CLUSTER_PORT
-            )
-            _ensure_card_unsigned_configmap(
-                kube=kube,
-                name=name,
-                namespace=namespace,
-                service_port=service_port,
-                description=f"Agent '{name}' deployed from UI.",
-                skill_names=final_skills or [],
-            )
-
-        # Create workload based on workloadType
-        if final_workload_type == WORKLOAD_TYPE_DEPLOYMENT:
-            workload_manifest = _build_deployment_manifest(
-                request=agent_request,
-                image=container_image,
-                shipwright_build_name=name,
-                local_skills=build_local_skills,
-                ext_init_containers=build_ext_init_containers,
-                ext_volumes=build_ext_volumes,
-                ext_volume_mounts=build_ext_volume_mounts,
-                ext_skill_paths=build_ext_skill_paths,
-            )
-            # Add additional labels from Build
-            workload_manifest["metadata"]["labels"].update(
-                {k: v for k, v in build_labels.items() if k.startswith("rossoctl.io/")}
-            )
-            # Also update pod template labels
-            workload_manifest["spec"]["template"]["metadata"]["labels"].update(
-                {k: v for k, v in build_labels.items() if k.startswith("rossoctl.io/")}
-            )
-            kube.create_deployment(namespace=namespace, body=workload_manifest)
-            logger.info(
-                f"Created Deployment '{name}' with image '{container_image}' in namespace '{namespace}'"
-            )
-        elif final_workload_type == WORKLOAD_TYPE_STATEFULSET:
-            workload_manifest = _build_statefulset_manifest(
-                request=agent_request,
-                image=container_image,
-                shipwright_build_name=name,
-                local_skills=build_local_skills,
-                ext_init_containers=build_ext_init_containers,
-                ext_volumes=build_ext_volumes,
-                ext_volume_mounts=build_ext_volume_mounts,
-                ext_skill_paths=build_ext_skill_paths,
-            )
-            # Add additional labels from Build
-            workload_manifest["metadata"]["labels"].update(
-                {k: v for k, v in build_labels.items() if k.startswith("rossoctl.io/")}
-            )
-            # Also update pod template labels
-            workload_manifest["spec"]["template"]["metadata"]["labels"].update(
-                {k: v for k, v in build_labels.items() if k.startswith("rossoctl.io/")}
-            )
-            kube.create_statefulset(namespace=namespace, body=workload_manifest)
-            logger.info(
-                f"Created StatefulSet '{name}' with image '{container_image}' in namespace '{namespace}'"
-            )
-        elif final_workload_type == WORKLOAD_TYPE_JOB:
-            workload_manifest = _build_job_manifest(
-                request=agent_request,
-                image=container_image,
-                shipwright_build_name=name,
-                local_skills=build_local_skills,
-                ext_init_containers=build_ext_init_containers,
-                ext_volumes=build_ext_volumes,
-                ext_volume_mounts=build_ext_volume_mounts,
-                ext_skill_paths=build_ext_skill_paths,
-            )
-            # Add additional labels from Build
-            workload_manifest["metadata"]["labels"].update(
-                {k: v for k, v in build_labels.items() if k.startswith("rossoctl.io/")}
-            )
-            # Also update pod template labels
-            workload_manifest["spec"]["template"]["metadata"]["labels"].update(
-                {k: v for k, v in build_labels.items() if k.startswith("rossoctl.io/")}
-            )
-            kube.create_job(namespace=namespace, body=workload_manifest)
-            logger.info(
-                f"Created Job '{name}' with image '{container_image}' in namespace '{namespace}'"
-            )
-        elif final_workload_type == WORKLOAD_TYPE_SANDBOX:
-            sandbox_manifest = _build_sandbox_manifest(
-                request=agent_request,
-                image=container_image,
-                shipwright_build_name=name,
-                local_skills=build_local_skills,
-                ext_init_containers=build_ext_init_containers,
-                ext_volumes=build_ext_volumes,
-                ext_volume_mounts=build_ext_volume_mounts,
-                ext_skill_paths=build_ext_skill_paths,
-            )
-            rossoctl_build_labels = {
-                k: v
-                for k, v in build_labels.items()
-                if k.startswith(settings.rossoctl_label_prefix)
-            }
-            sandbox_manifest["metadata"]["labels"].update(rossoctl_build_labels)
-            sandbox_manifest["spec"]["podTemplate"]["metadata"]["labels"].update(
-                rossoctl_build_labels
-            )
-            kube.create_sandbox(namespace=namespace, body=sandbox_manifest)
-            logger.info(f"Created Sandbox '{name}' in namespace '{namespace}' from build")
-
-        # Create Service via the shared _create_or_replace_service helper
-        # (skips only for Job workloads).
-        service_manifest = _build_service_manifest(agent_request)
-        # Carry forward build-time rossoctl.io/* labels onto the Service so
-        # downstream label-based selectors / queries match. Use
-        # settings.rossoctl_label_prefix (the project-wide constant) instead
-        # of the literal "rossoctl.io/" so CodeQL's URL-substring rule
-        # doesn't pattern-match the literal — see line 3626 above for the
-        # same idiom.
-        service_manifest["metadata"]["labels"].update(
-            {k: v for k, v in build_labels.items() if k.startswith(settings.rossoctl_label_prefix)}
-        )
-        _create_or_replace_service(kube, namespace, name, service_manifest, final_workload_type)
-
-        # Create AgentRuntime CR so the per-agent AuthBridge config is applied.
-        # Sandbox is included (targetRef -> agents.x-k8s.io Sandbox); only Job is
-        # excluded. Agents only — tools don't need sidecar injection.
-        resource_type = build_labels.get(ROSSOCTL_TYPE_LABEL, RESOURCE_TYPE_AGENT)
-        if (
-            _agentruntime_supported_workload(final_workload_type)
-            and resource_type == RESOURCE_TYPE_AGENT
-        ):
-            _ensure_agentruntime(
-                kube=kube,
-                name=name,
-                namespace=namespace,
-                workload_type=final_workload_type,
-                auth_bridge_mode=final_auth_bridge_mode,
-                mtls_mode=final_mtls_mode,
-                tls_bridge_enabled=final_tls_bridge_enabled,
-            )
-
-        message = f"Agent '{name}' deployed as {final_workload_type} with image '{output_image}'."
-
-        # Step 4: Create HTTPRoute/Route if requested (not applicable for Jobs or Sandboxes)
-        if final_create_route and final_workload_type not in (
-            WORKLOAD_TYPE_JOB,
-            WORKLOAD_TYPE_SANDBOX,
-        ):
-            service_port = select_route_port(
-                final_service_ports,
-                default_port=DEFAULT_OFF_CLUSTER_PORT,
-            )
-            create_route_for_agent_or_tool(
-                kube=kube,
-                name=name,
-                namespace=namespace,
-                service_name=name,
-                service_port=service_port,
-            )
-            message += " HTTPRoute/Route created for external access."
-
-        return CreateAgentResponse(
-            success=True,
-            name=name,
-            namespace=namespace,
-            message=message,
-        )
-
-    except ApiException as e:
-        if e.status == 409:
-            raise HTTPException(
-                status_code=409,
-                detail=f"Agent '{name}' already exists in namespace '{namespace}'",
-            )
-        logger.error(f"Failed to finalize build: {e}")
-        raise HTTPException(status_code=e.status, detail=str(e.reason))
-
-
-# New models for env parsing
-class ParseEnvRequest(BaseModel):
-    """Request to parse .env file content."""
-
-    content: str
-
-
-class ParseEnvResponse(BaseModel):
-    """Response with parsed environment variables."""
-
-    envVars: List[Dict[str, Any]]
-    warnings: Optional[List[str]] = None
-
-
-class FetchEnvUrlRequest(BaseModel):
-    """Request to fetch .env file from URL."""
-
-    url: str
-
-
-class FetchEnvUrlResponse(BaseModel):
-    """Response with fetched .env file content."""
-
-    content: str
-    url: str
-
-
-# Blocked IP ranges for SSRF protection
-BLOCKED_IP_RANGES = [
-    ipaddress.ip_network("10.0.0.0/8"),
-    ipaddress.ip_network("172.16.0.0/12"),
-    ipaddress.ip_network("192.168.0.0/16"),
-    ipaddress.ip_network("127.0.0.0/8"),
-    ipaddress.ip_network("169.254.0.0/16"),
+)(create_agent)
+router.include_router(finalize_router)
+router.include_router(env_router)
+router.include_router(authbridge_router)
+
+# Names re-exported for backwards compatibility (imported by app.routers.tools,
+# app.services.reconciliation, app.services.agent_env_defaults and the tests).
+__all__ = [
+    # router + endpoints owned by this module
+    "router",
+    "list_agents",
+    "get_agent",
+    "get_agent_route_status",
+    "delete_agent",
+    # endpoints defined in submodules
+    "create_agent",
+    "fetch_env_from_url",
+    "finalize_shipwright_build",
+    "get_shipwright_build_info",
+    "get_shipwright_build_status",
+    "get_shipwright_buildrun_status",
+    "list_agent_shipwright_builds",
+    "list_build_strategies",
+    "list_migratable_agents",
+    "migrate_agent",
+    "migrate_all_agents",
+    "parse_env_file",
+    "trigger_shipwright_buildrun",
+    # models
+    "AgentShipwrightBuildInfoResponse",
+    "ConfigMapKeyRef",
+    "ContextAttachment",
+    "CreateAgentRequest",
+    "CreateAgentResponse",
+    "EnvVar",
+    "EnvVarSource",
+    "FetchEnvUrlRequest",
+    "FetchEnvUrlResponse",
+    "FinalizeShipwrightBuildRequest",
+    "ListMigratableAgentsResponse",
+    "MigratableAgentInfo",
+    "MigrateAgentRequest",
+    "MigrateAgentResponse",
+    "OutboundRoute",
+    "ParseEnvRequest",
+    "ParseEnvResponse",
+    "PersistentStorageConfig",
+    "SecretKeyRef",
+    "ServicePort",
+    # helpers
+    "build_container_resources",
+    "is_ip_blocked",
+    "BLOCKED_IP_RANGES",
+    "_agentruntime_supported_workload",
+    "_build_agent_shipwright_build_manifest",
+    "_build_agent_shipwright_buildrun_manifest",
+    "_build_agentruntime_manifest",
+    "_build_authbridge_runtime_yaml",
+    "_build_authbridge_runtime_yaml_fallback",
+    "_build_common_annotations",
+    "_build_common_labels",
+    "_build_deployment_from_agent_crd",
+    "_build_deployment_manifest",
+    "_build_env_vars",
+    "_build_fetcher_scripts_data",
+    "_build_job_manifest",
+    "_build_sandbox_manifest",
+    "_build_selector_labels",
+    "_build_service_from_agent_crd",
+    "_build_service_manifest",
+    "_build_statefulset_manifest",
+    "_create_or_replace_service",
+    "_ensure_agentruntime",
+    "_resolve_context_mounts",
+    "_ensure_authbridge_configmaps",
+    "_ensure_authbridge_scc_rolebinding",
+    "_ensure_authproxy_routes",
+    "_ensure_card_unsigned_configmap",
+    "_ensure_fetcher_scripts_cm",
+    "_extract_labels",
+    "_fetch_authbridge_json",
+    "_format_timestamp",
+    "_get_authbridge_runtime_yaml",
+    "_get_deployment_description",
+    "_get_external_skill_data",
+    "_get_job_description",
+    "_get_job_status",
+    "_get_linked_skill_mounts",
+    "_get_sandbox_description",
+    "_get_service_endpoints",
+    "_get_statefulset_description",
+    "_is_deployment_ready",
+    "_is_sandbox_ready",
+    "_is_skill_external",
+    "_is_statefulset_ready",
+    "_load_agent_skill_summaries",
+    # models/constants re-exported for tests
+    "ResourceLabels",
+    "ShipwrightBuildConfig",
+    "WORKLOAD_TYPE_DEPLOYMENT",
+    "WORKLOAD_TYPE_JOB",
+    "WORKLOAD_TYPE_SANDBOX",
+    "WORKLOAD_TYPE_STATEFULSET",
 ]
-
-
-def is_ip_blocked(ip_str: str) -> bool:
-    """Check if IP is in blocked range for SSRF protection."""
-    try:
-        ip = ipaddress.ip_address(ip_str)
-        return any(ip in network for network in BLOCKED_IP_RANGES)
-    except ValueError:
-        return False
-
-
-@router.post(
-    "/parse-env",
-    response_model=ParseEnvResponse,
-    dependencies=[Depends(require_roles(ROLE_OPERATOR))],
-)
-async def parse_env_file(request: ParseEnvRequest) -> ParseEnvResponse:
-    """
-    Parse .env file content and return structured environment variables.
-    Supports:
-    - Standard KEY=value format
-    - Extended JSON format for secretKeyRef and configMapKeyRef
-
-    Example extended format:
-    SECRET_KEY='{"valueFrom": {"secretKeyRef": {"name": "openai-secret", "key": "apikey"}}}'
-    """
-    env_vars = []
-    warnings = []
-
-    lines = request.content.strip().split("\n")
-
-    for line_num, line in enumerate(lines, 1):
-        # Skip empty lines and comments
-        line = line.strip()
-        if not line or line.startswith("#"):
-            continue
-
-        # Parse KEY=VALUE
-        if "=" not in line:
-            warnings.append(f"Line {line_num}: Invalid format, missing '='")
-            continue
-
-        key, value = line.split("=", 1)
-        key = key.strip()
-        value = value.strip()
-
-        # Validate environment variable name
-        env_var_pattern = r"^[A-Za-z_][A-Za-z0-9_]*$"
-        if not re.match(env_var_pattern, key):
-            warnings.append(
-                f"Line {line_num}: Invalid variable name '{key}'. "
-                "Name must start with a letter or underscore and contain only "
-                "letters, digits, and underscores."
-            )
-            continue
-
-        # Remove quotes if present
-        if (value.startswith('"') and value.endswith('"')) or (
-            value.startswith("'") and value.endswith("'")
-        ):
-            value = value[1:-1]
-
-        # Try to parse as JSON (for extended format)
-        if value.startswith("{") and value.endswith("}"):
-            try:
-                parsed = json.loads(value)
-                if "valueFrom" in parsed:
-                    env_var = {"name": key, "valueFrom": parsed["valueFrom"]}
-                    env_vars.append(env_var)
-                    continue
-                else:
-                    # It's valid JSON but not our expected format, treat as string
-                    warnings.append(
-                        f"Line {line_num}: JSON value without 'valueFrom' key, treating as string"
-                    )
-            except json.JSONDecodeError as e:
-                warnings.append(f"Line {line_num}: Invalid JSON in value: {str(e)}")
-
-        # Standard value
-        env_vars.append({"name": key, "value": value})
-
-    return ParseEnvResponse(envVars=env_vars, warnings=warnings if warnings else None)
-
-
-@router.post(
-    "/fetch-env-url",
-    response_model=FetchEnvUrlResponse,
-    dependencies=[Depends(require_roles(ROLE_OPERATOR))],
-)
-async def fetch_env_from_url(request: FetchEnvUrlRequest) -> FetchEnvUrlResponse:
-    """
-    Fetch .env file content from a remote URL.
-    Supports HTTP/HTTPS URLs with security validations to prevent SSRF attacks.
-
-    Example URLs:
-    - https://raw.githubusercontent.com/rossoctl/examples/main/a2a/git_issue_agent/.env.openai
-    - https://example.com/config/.env
-    """
-    import os
-    import ssl
-    from pathlib import Path
-
-    logger.info(f"Fetching .env file from URL: {request.url}")
-
-    # Log SSL/Certificate configuration
-    logger.info(f"SSL_CERT_FILE env: {os.environ.get('SSL_CERT_FILE', 'NOT SET')}")
-    logger.info(f"REQUESTS_CA_BUNDLE env: {os.environ.get('REQUESTS_CA_BUNDLE', 'NOT SET')}")
-    logger.info(f"Default SSL context: {ssl.get_default_verify_paths()}")
-
-    # Check if cert files exist
-    cert_paths = [
-        "/etc/ssl/certs/ca-certificates.crt",
-        "/etc/ssl/certs/ca-bundle.crt",
-        "/usr/local/share/ca-certificates/",
-    ]
-    for cert_path in cert_paths:
-        exists = (
-            Path(cert_path).exists() if cert_path.endswith(".crt") else Path(cert_path).is_dir()
-        )
-        logger.info(f"Certificate path {cert_path}: {'EXISTS' if exists else 'NOT FOUND'}")
-
-    # Security validation - only allow http/https
-    parsed_url = urlparse(request.url)
-    if parsed_url.scheme not in ["http", "https"]:
-        raise HTTPException(status_code=400, detail="Only HTTP/HTTPS URLs are supported")
-
-    # Validate hostname exists
-    if not parsed_url.hostname:
-        raise HTTPException(status_code=400, detail="Invalid URL: hostname not found")
-
-    # Prevent SSRF attacks - block private IPs
-    try:
-        ip = socket.gethostbyname(parsed_url.hostname)
-        logger.debug(f"Resolved {parsed_url.hostname} to {ip}")
-        if is_ip_blocked(ip):
-            logger.warning(f"Blocked private IP address: {ip}")
-            raise HTTPException(
-                status_code=400, detail="Private IP addresses are not allowed for security reasons"
-            )
-    except socket.gaierror as e:
-        # Domain can't be resolved - log but let httpx handle it
-        logger.warning(f"Could not resolve hostname {parsed_url.hostname}: {e}")
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.warning(f"Error checking IP for {parsed_url.hostname}: {e}")
-
-    # Fetch content with timeout
-    try:
-        # Explicitly use system CA bundle instead of Kubernetes service account CA
-        # Kubernetes sets SSL_CERT_FILE to /var/run/secrets/kubernetes.io/serviceaccount/ca.crt
-        # which doesn't include public CAs like GitHub. We need to explicitly point to system CAs.
-        ca_bundle_path = "/etc/ssl/certs/ca-certificates.crt"
-        if not Path(ca_bundle_path).exists():
-            # Fallback to alternative paths
-            for fallback in ["/etc/ssl/certs/ca-bundle.crt", "/etc/pki/tls/certs/ca-bundle.crt"]:
-                if Path(fallback).exists():
-                    ca_bundle_path = fallback
-                    break
-
-        logger.info(f"Using CA bundle: {ca_bundle_path}")
-
-        # Create SSL context with system certificates
-        ssl_context = ssl.create_default_context(cafile=ca_bundle_path)
-
-        async with httpx.AsyncClient(
-            timeout=10.0, follow_redirects=True, verify=ssl_context
-        ) as client:
-            logger.debug(f"Making HTTP request to {request.url}")
-            response = await client.get(request.url)
-            response.raise_for_status()
-
-            logger.info(f"Successfully fetched URL, content length: {len(response.text)} bytes")
-
-            # Validate content isn't too large (max 1MB)
-            content = response.text
-            if len(content) > 1024 * 1024:
-                raise HTTPException(status_code=413, detail="File content too large (max 1MB)")
-
-            return FetchEnvUrlResponse(content=content, url=request.url)
-    except httpx.TimeoutException as e:
-        logger.error(f"Timeout fetching URL {request.url}: {e}")
-        raise HTTPException(status_code=504, detail="Request timeout while fetching URL")
-    except httpx.HTTPStatusError as e:
-        logger.error(f"HTTP error fetching URL {request.url}: {e.response.status_code}")
-        raise HTTPException(
-            status_code=e.response.status_code,
-            detail=f"Failed to fetch URL: {e.response.status_code} {e.response.reason_phrase}",
-        )
-    except httpx.HTTPError as e:
-        logger.error(f"HTTP error fetching URL {request.url}: {str(e)}")
-        raise HTTPException(status_code=502, detail=f"Failed to fetch URL: {str(e)}")
-    except Exception as e:
-        logger.error(f"Unexpected error fetching URL {request.url}: {str(e)}")
-        raise HTTPException(status_code=500, detail=f"Unexpected error: {str(e)}")
-
-
-_K8S_NAME_RE = re.compile(r"^[a-z0-9]([a-z0-9-]{0,61}[a-z0-9])?$")
-
-
-if settings.rossoctl_feature_flag_authbridge_api:
-
-    @router.get(
-        "/{namespace}/{name}/identity-config", dependencies=[Depends(require_roles(ROLE_OPERATOR))]
-    )
-    async def get_agent_identity_config(
-        namespace: str,
-        name: str,
-        kube: KubernetesService = Depends(get_kubernetes_service),
-    ) -> dict:
-        """
-        Fetch the AuthBridge configuration for an Agent.
-        """
-
-        namespace = sanitize_log(namespace)
-        name = sanitize_log(name)
-
-        try:
-            addresses = _get_service_endpoints(kube=kube, namespace=namespace, name=name)
-        except ApiException as e:
-            raise HTTPException(status_code=502, detail=e.reason)
-
-        attempts = 0
-        for address in addresses:
-            attempts += 1
-            # AuthBridge serves config and status on port 9093
-            url = f"http://{address}:9093/config"
-            try:
-                data = await _fetch_authbridge_json(url)
-                data["AuthBridge"] = True
-                return data
-            except Exception:
-                # It isn't an error for an endpoint to be unreachable, only for all pods to be unreachable
-                logger.info("Failed to talk to url %s; skipping", url, exc_info=True)
-
-        if attempts == 0:
-            raise HTTPException(status_code=404, detail=f"{name} not found")
-
-        logger.info("Could not invoke any AuthBridge endpoints for %s/%s", namespace, name)
-        # We return HTTP 200 if no pods respond - this might be a valid agent w/o AuthBridge
-        return {"AuthBridge": False}
-
-    @router.put(
-        "/{namespace}/{name}/identity-config", dependencies=[Depends(require_roles(ROLE_OPERATOR))]
-    )
-    async def put_agent_identity_config(
-        namespace: str,
-        name: str,
-        new_authbridge_config_yaml: str = Body(media_type="text/plain"),
-        kube: KubernetesService = Depends(get_kubernetes_service),
-    ) -> dict:
-        """
-        Set the AuthBridge configuration for an Agent.
-        """
-
-        if not _K8S_NAME_RE.fullmatch(namespace):
-            raise HTTPException(status_code=400, detail="Invalid namespace")
-        if not _K8S_NAME_RE.fullmatch(name):
-            raise HTTPException(status_code=400, detail="Invalid name")
-
-        try:
-            yaml.safe_load(new_authbridge_config_yaml)
-        except yaml.YAMLError as e:
-            raise HTTPException(status_code=400, detail=f"Invalid YAML: {e}") from e
-
-        kube.upsert_configmap(
-            namespace=namespace,
-            name=f"authbridge-config-{name}",
-            data={"config.yaml": new_authbridge_config_yaml},
-        )
-
-        return {"status": "ok"}
-
-    @router.get(
-        "/{namespace}/{name}/identity-status", dependencies=[Depends(require_roles(ROLE_OPERATOR))]
-    )
-    async def get_agent_identity_status(
-        namespace: str,
-        name: str,
-        kube: KubernetesService = Depends(get_kubernetes_service),
-    ) -> dict:
-        """
-        Fetch the AuthBridge statistics and status for an Agent.
-        """
-
-        namespace = sanitize_log(namespace)
-        name = sanitize_log(name)
-
-        try:
-            addresses = _get_service_endpoints(kube=kube, namespace=namespace, name=name)
-        except ApiException as e:
-            raise HTTPException(status_code=502, detail=e.reason)
-
-        attempts = 0
-        for address in addresses:
-            attempts += 1
-            # AuthBridge serves config and status on port 9093
-            url = f"http://{address}:9093/stats"
-            try:
-                data = await _fetch_authbridge_json(url)
-                data["AuthBridge"] = True
-                return data
-            except Exception:
-                # It isn't an error for an endpoint to be unreachable, only for all pods to be unreachable
-                logger.info("Failed to talk to url %s; skipping", url, exc_info=True)
-
-        if attempts == 0:
-            raise HTTPException(status_code=404, detail=f"{name} not found")
-
-        logger.info("Could not invoke any AuthBridge endpoints for %s/%s", namespace, name)
-        # We return HTTP 200 if no pods respond - this might be a valid agent w/o AuthBridge
-        return {"AuthBridge": False}
-
-
-def _get_service_endpoints(kube: KubernetesService, namespace: str, name: str) -> List[str]:
-    """
-    Get addresses for a K8s service
-    """
-
-    addresses: list[str] = []
-    endpoint_slices = kube.get_endpoint_slices(namespace=namespace, name=name)
-
-    for endpoint_slice in endpoint_slices.get("items", []):
-        for endpoint in endpoint_slice.get("endpoints", []):
-            for address in endpoint.get("addresses", []):
-                addresses.append(address)
-
-    return addresses
-
-
-async def _fetch_authbridge_json(url: str) -> dict:
-    """
-    Fetch JSON from an AuthBridge sidecar endpoint.
-
-    Raises on HTTP errors, oversized responses, or non-dict payloads.
-    """
-
-    async with httpx.AsyncClient(timeout=10.0, follow_redirects=True) as client:
-        logger.debug("Making HTTP request to %s", url)
-        response = await client.get(url)
-        response.raise_for_status()
-
-        content = response.text
-        if len(content) > 1024 * 1024:
-            raise HTTPException(status_code=502, detail="File content too large (max 1MB)")
-
-        data = json.loads(content)
-        if not isinstance(data, dict):
-            raise HTTPException(status_code=502, detail="File content not AuthBridge JSON")
-
-        return data
